switched to joyride foundation, does really help solving the existing issues but a lot more light weight
