+++ cs130_LL1/static/css/foundation.joyride.css
@@ -1,0 +1,20 @@
+[data-joyride] {
+  display: none; }
+
+.tooltip.joyride,
+.reveal.joyride {
+  font-size: 100%;
+  min-width: 20rem;
+  padding: 0.5em; }
+  .tooltip.joyride .close,
+  .reveal.joyride .close {
+    float: right; }
+  .tooltip.joyride .joyride-buttons,
+  .reveal.joyride .joyride-buttons {
+    text-align: center; }
+    .tooltip.joyride .joyride-buttons button,
+    .reveal.joyride .joyride-buttons button {
+      margin: 0 0.5em; }
+
+[data-whatinput='mouse'] .joyride {
+  outline: 0; }
