--- cs130_LL1/static/js/solo.joyride.js
@@ -1,3439 +1,0 @@
-!function ($) {
-
-  "use strict";
-
-  var FOUNDATION_VERSION = '6.3.1';
-
-  // Global Foundation object
-  // This is attached to the window, or used as a module for AMD/Browserify
-  var Foundation = {
-    version: FOUNDATION_VERSION,
-
-    /**
-     * Stores initialized plugins.
-     */
-    _plugins: {},
-
-    /**
-     * Stores generated unique ids for plugin instances
-     */
-    _uuids: [],
-
-    /**
-     * Returns a boolean for RTL support
-     */
-    rtl: function () {
-      return $('html').attr('dir') === 'rtl';
-    },
-    /**
-     * Defines a Foundation plugin, adding it to the `Foundation` namespace and the list of plugins to initialize when reflowing.
-     * @param {Object} plugin - The constructor of the plugin.
-     */
-    plugin: function (plugin, name) {
-      // Object key to use when adding to global Foundation object
-      // Examples: Foundation.Reveal, Foundation.OffCanvas
-      var className = name || functionName(plugin);
-      // Object key to use when storing the plugin, also used to create the identifying data attribute for the plugin
-      // Examples: data-reveal, data-off-canvas
-      var attrName = hyphenate(className);
-
-      // Add to the Foundation object and the plugins list (for reflowing)
-      this._plugins[attrName] = this[className] = plugin;
-    },
-    /**
-     * @function
-     * Populates the _uuids array with pointers to each individual plugin instance.
-     * Adds the `zfPlugin` data-attribute to programmatically created plugins to allow use of $(selector).foundation(method) calls.
-     * Also fires the initialization event for each plugin, consolidating repetitive code.
-     * @param {Object} plugin - an instance of a plugin, usually `this` in context.
-     * @param {String} name - the name of the plugin, passed as a camelCased string.
-     * @fires Plugin#init
-     */
-    registerPlugin: function (plugin, name) {
-      var pluginName = name ? hyphenate(name) : functionName(plugin.constructor).toLowerCase();
-      plugin.uuid = this.GetYoDigits(6, pluginName);
-
-      if (!plugin.$element.attr('data-' + pluginName)) {
-        plugin.$element.attr('data-' + pluginName, plugin.uuid);
-      }
-      if (!plugin.$element.data('zfPlugin')) {
-        plugin.$element.data('zfPlugin', plugin);
-      }
-      /**
-       * Fires when the plugin has initialized.
-       * @event Plugin#init
-       */
-      plugin.$element.trigger('init.zf.' + pluginName);
-
-      this._uuids.push(plugin.uuid);
-
-      return;
-    },
-    /**
-     * @function
-     * Removes the plugins uuid from the _uuids array.
-     * Removes the zfPlugin data attribute, as well as the data-plugin-name attribute.
-     * Also fires the destroyed event for the plugin, consolidating repetitive code.
-     * @param {Object} plugin - an instance of a plugin, usually `this` in context.
-     * @fires Plugin#destroyed
-     */
-    unregisterPlugin: function (plugin) {
-      var pluginName = hyphenate(functionName(plugin.$element.data('zfPlugin').constructor));
-
-      this._uuids.splice(this._uuids.indexOf(plugin.uuid), 1);
-      plugin.$element.removeAttr('data-' + pluginName).removeData('zfPlugin')
-      /**
-       * Fires when the plugin has been destroyed.
-       * @event Plugin#destroyed
-       */
-      .trigger('destroyed.zf.' + pluginName);
-      for (var prop in plugin) {
-        plugin[prop] = null; //clean up script to prep for garbage collection.
-      }
-      return;
-    },
-
-    /**
-     * @function
-     * Causes one or more active plugins to re-initialize, resetting event listeners, recalculating positions, etc.
-     * @param {String} plugins - optional string of an individual plugin key, attained by calling `$(element).data('pluginName')`, or string of a plugin class i.e. `'dropdown'`
-     * @default If no argument is passed, reflow all currently active plugins.
-     */
-    reInit: function (plugins) {
-      var isJQ = plugins instanceof $;
-      try {
-        if (isJQ) {
-          plugins.each(function () {
-            $(this).data('zfPlugin')._init();
-          });
-        } else {
-          var type = typeof plugins,
-              _this = this,
-              fns = {
-            'object': function (plgs) {
-              plgs.forEach(function (p) {
-                p = hyphenate(p);
-                $('[data-' + p + ']').foundation('_init');
-              });
-            },
-            'string': function () {
-              plugins = hyphenate(plugins);
-              $('[data-' + plugins + ']').foundation('_init');
-            },
-            'undefined': function () {
-              this['object'](Object.keys(_this._plugins));
-            }
-          };
-          fns[type](plugins);
-        }
-      } catch (err) {
-        console.error(err);
-      } finally {
-        return plugins;
-      }
-    },
-
-    /**
-     * returns a random base-36 uid with namespacing
-     * @function
-     * @param {Number} length - number of random base-36 digits desired. Increase for more random strings.
-     * @param {String} namespace - name of plugin to be incorporated in uid, optional.
-     * @default {String} '' - if no plugin name is provided, nothing is appended to the uid.
-     * @returns {String} - unique id
-     */
-    GetYoDigits: function (length, namespace) {
-      length = length || 6;
-      return Math.round(Math.pow(36, length + 1) - Math.random() * Math.pow(36, length)).toString(36).slice(1) + (namespace ? '-' + namespace : '');
-    },
-    /**
-     * Initialize plugins on any elements within `elem` (and `elem` itself) that aren't already initialized.
-     * @param {Object} elem - jQuery object containing the element to check inside. Also checks the element itself, unless it's the `document` object.
-     * @param {String|Array} plugins - A list of plugins to initialize. Leave this out to initialize everything.
-     */
-    reflow: function (elem, plugins) {
-
-      // If plugins is undefined, just grab everything
-      if (typeof plugins === 'undefined') {
-        plugins = Object.keys(this._plugins);
-      }
-      // If plugins is a string, convert it to an array with one item
-      else if (typeof plugins === 'string') {
-          plugins = [plugins];
-        }
-
-      var _this = this;
-
-      // Iterate through each plugin
-      $.each(plugins, function (i, name) {
-        // Get the current plugin
-        var plugin = _this._plugins[name];
-
-        // Localize the search to all elements inside elem, as well as elem itself, unless elem === document
-        var $elem = $(elem).find('[data-' + name + ']').addBack('[data-' + name + ']');
-
-        // For each plugin found, initialize it
-        $elem.each(function () {
-          var $el = $(this),
-              opts = {};
-          // Don't double-dip on plugins
-          if ($el.data('zfPlugin')) {
-            console.warn("Tried to initialize " + name + " on an element that already has a Foundation plugin.");
-            return;
-          }
-
-          if ($el.attr('data-options')) {
-            var thing = $el.attr('data-options').split(';').forEach(function (e, i) {
-              var opt = e.split(':').map(function (el) {
-                return el.trim();
-              });
-              if (opt[0]) opts[opt[0]] = parseValue(opt[1]);
-            });
-          }
-          try {
-            $el.data('zfPlugin', new plugin($(this), opts));
-          } catch (er) {
-            console.error(er);
-          } finally {
-            return;
-          }
-        });
-      });
-    },
-    getFnName: functionName,
-    transitionend: function ($elem) {
-      var transitions = {
-        'transition': 'transitionend',
-        'WebkitTransition': 'webkitTransitionEnd',
-        'MozTransition': 'transitionend',
-        'OTransition': 'otransitionend'
-      };
-      var elem = document.createElement('div'),
-          end;
-
-      for (var t in transitions) {
-        if (typeof elem.style[t] !== 'undefined') {
-          end = transitions[t];
-        }
-      }
-      if (end) {
-        return end;
-      } else {
-        end = setTimeout(function () {
-          $elem.triggerHandler('transitionend', [$elem]);
-        }, 1);
-        return 'transitionend';
-      }
-    }
-  };
-
-  Foundation.util = {
-    /**
-     * Function for applying a debounce effect to a function call.
-     * @function
-     * @param {Function} func - Function to be called at end of timeout.
-     * @param {Number} delay - Time in ms to delay the call of `func`.
-     * @returns function
-     */
-    throttle: function (func, delay) {
-      var timer = null;
-
-      return function () {
-        var context = this,
-            args = arguments;
-
-        if (timer === null) {
-          timer = setTimeout(function () {
-            func.apply(context, args);
-            timer = null;
-          }, delay);
-        }
-      };
-    }
-  };
-
-  // TODO: consider not making this a jQuery function
-  // TODO: need way to reflow vs. re-initialize
-  /**
-   * The Foundation jQuery method.
-   * @param {String|Array} method - An action to perform on the current jQuery object.
-   */
-  var foundation = function (method) {
-    var type = typeof method,
-        $meta = $('meta.foundation-mq'),
-        $noJS = $('.no-js');
-
-    if (!$meta.length) {
-      $('<meta class="foundation-mq">').appendTo(document.head);
-    }
-    if ($noJS.length) {
-      $noJS.removeClass('no-js');
-    }
-
-    if (type === 'undefined') {
-      //needs to initialize the Foundation object, or an individual plugin.
-      Foundation.MediaQuery._init();
-      Foundation.reflow(this);
-    } else if (type === 'string') {
-      //an individual method to invoke on a plugin or group of plugins
-      var args = Array.prototype.slice.call(arguments, 1); //collect all the arguments, if necessary
-      var plugClass = this.data('zfPlugin'); //determine the class of plugin
-
-      if (plugClass !== undefined && plugClass[method] !== undefined) {
-        //make sure both the class and method exist
-        if (this.length === 1) {
-          //if there's only one, call it directly.
-          plugClass[method].apply(plugClass, args);
-        } else {
-          this.each(function (i, el) {
-            //otherwise loop through the jQuery collection and invoke the method on each
-            plugClass[method].apply($(el).data('zfPlugin'), args);
-          });
-        }
-      } else {
-        //error for no class or no method
-        throw new ReferenceError("We're sorry, '" + method + "' is not an available method for " + (plugClass ? functionName(plugClass) : 'this element') + '.');
-      }
-    } else {
-      //error for invalid argument type
-      throw new TypeError('We\'re sorry, ' + type + ' is not a valid parameter. You must use a string representing the method you wish to invoke.');
-    }
-    return this;
-  };
-
-  window.Foundation = Foundation;
-  $.fn.foundation = foundation;
-
-  // Polyfill for requestAnimationFrame
-  (function () {
-    if (!Date.now || !window.Date.now) window.Date.now = Date.now = function () {
-      return new Date().getTime();
-    };
-
-    var vendors = ['webkit', 'moz'];
-    for (var i = 0; i < vendors.length && !window.requestAnimationFrame; ++i) {
-      var vp = vendors[i];
-      window.requestAnimationFrame = window[vp + 'RequestAnimationFrame'];
-      window.cancelAnimationFrame = window[vp + 'CancelAnimationFrame'] || window[vp + 'CancelRequestAnimationFrame'];
-    }
-    if (/iP(ad|hone|od).*OS 6/.test(window.navigator.userAgent) || !window.requestAnimationFrame || !window.cancelAnimationFrame) {
-      var lastTime = 0;
-      window.requestAnimationFrame = function (callback) {
-        var now = Date.now();
-        var nextTime = Math.max(lastTime + 16, now);
-        return setTimeout(function () {
-          callback(lastTime = nextTime);
-        }, nextTime - now);
-      };
-      window.cancelAnimationFrame = clearTimeout;
-    }
-    /**
-     * Polyfill for performance.now, required by rAF
-     */
-    if (!window.performance || !window.performance.now) {
-      window.performance = {
-        start: Date.now(),
-        now: function () {
-          return Date.now() - this.start;
-        }
-      };
-    }
-  })();
-  if (!Function.prototype.bind) {
-    Function.prototype.bind = function (oThis) {
-      if (typeof this !== 'function') {
-        // closest thing possible to the ECMAScript 5
-        // internal IsCallable function
-        throw new TypeError('Function.prototype.bind - what is trying to be bound is not callable');
-      }
-
-      var aArgs = Array.prototype.slice.call(arguments, 1),
-          fToBind = this,
-          fNOP = function () {},
-          fBound = function () {
-        return fToBind.apply(this instanceof fNOP ? this : oThis, aArgs.concat(Array.prototype.slice.call(arguments)));
-      };
-
-      if (this.prototype) {
-        // native functions don't have a prototype
-        fNOP.prototype = this.prototype;
-      }
-      fBound.prototype = new fNOP();
-
-      return fBound;
-    };
-  }
-  // Polyfill to get the name of a function in IE9
-  function functionName(fn) {
-    if (Function.prototype.name === undefined) {
-      var funcNameRegex = /function\s([^(]{1,})\(/;
-      var results = funcNameRegex.exec(fn.toString());
-      return results && results.length > 1 ? results[1].trim() : "";
-    } else if (fn.prototype === undefined) {
-      return fn.constructor.name;
-    } else {
-      return fn.prototype.constructor.name;
-    }
-  }
-  function parseValue(str) {
-    if ('true' === str) return true;else if ('false' === str) return false;else if (!isNaN(str * 1)) return parseFloat(str);
-    return str;
-  }
-  // Convert PascalCase to kebab-case
-  // Thank you: http://stackoverflow.com/a/8955580
-  function hyphenate(str) {
-    return str.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase();
-  }
-}(jQuery);
-'use strict';
-
-!function ($) {
-
-  Foundation.Box = {
-    ImNotTouchingYou: ImNotTouchingYou,
-    GetDimensions: GetDimensions,
-    GetOffsets: GetOffsets
-  };
-
-  /**
-   * Compares the dimensions of an element to a container and determines collision events with container.
-   * @function
-   * @param {jQuery} element - jQuery object to test for collisions.
-   * @param {jQuery} parent - jQuery object to use as bounding container.
-   * @param {Boolean} lrOnly - set to true to check left and right values only.
-   * @param {Boolean} tbOnly - set to true to check top and bottom values only.
-   * @default if no parent object passed, detects collisions with `window`.
-   * @returns {Boolean} - true if collision free, false if a collision in any direction.
-   */
-  function ImNotTouchingYou(element, parent, lrOnly, tbOnly) {
-    var eleDims = GetDimensions(element),
-        top,
-        bottom,
-        left,
-        right;
-
-    if (parent) {
-      var parDims = GetDimensions(parent);
-
-      bottom = eleDims.offset.top + eleDims.height <= parDims.height + parDims.offset.top;
-      top = eleDims.offset.top >= parDims.offset.top;
-      left = eleDims.offset.left >= parDims.offset.left;
-      right = eleDims.offset.left + eleDims.width <= parDims.width + parDims.offset.left;
-    } else {
-      bottom = eleDims.offset.top + eleDims.height <= eleDims.windowDims.height + eleDims.windowDims.offset.top;
-      top = eleDims.offset.top >= eleDims.windowDims.offset.top;
-      left = eleDims.offset.left >= eleDims.windowDims.offset.left;
-      right = eleDims.offset.left + eleDims.width <= eleDims.windowDims.width;
-    }
-
-    var allDirs = [bottom, top, left, right];
-
-    if (lrOnly) {
-      return left === right === true;
-    }
-
-    if (tbOnly) {
-      return top === bottom === true;
-    }
-
-    return allDirs.indexOf(false) === -1;
-  };
-
-  /**
-   * Uses native methods to return an object of dimension values.
-   * @function
-   * @param {jQuery || HTML} element - jQuery object or DOM element for which to get the dimensions. Can be any element other that document or window.
-   * @returns {Object} - nested object of integer pixel values
-   * TODO - if element is window, return only those values.
-   */
-  function GetDimensions(elem, test) {
-    elem = elem.length ? elem[0] : elem;
-
-    if (elem === window || elem === document) {
-      throw new Error("I'm sorry, Dave. I'm afraid I can't do that.");
-    }
-
-    var rect = elem.getBoundingClientRect(),
-        parRect = elem.parentNode.getBoundingClientRect(),
-        winRect = document.body.getBoundingClientRect(),
-        winY = window.pageYOffset,
-        winX = window.pageXOffset;
-
-    return {
-      width: rect.width,
-      height: rect.height,
-      offset: {
-        top: rect.top + winY,
-        left: rect.left + winX
-      },
-      parentDims: {
-        width: parRect.width,
-        height: parRect.height,
-        offset: {
-          top: parRect.top + winY,
-          left: parRect.left + winX
-        }
-      },
-      windowDims: {
-        width: winRect.width,
-        height: winRect.height,
-        offset: {
-          top: winY,
-          left: winX
-        }
-      }
-    };
-  }
-
-  /**
-   * Returns an object of top and left integer pixel values for dynamically rendered elements,
-   * such as: Tooltip, Reveal, and Dropdown
-   * @function
-   * @param {jQuery} element - jQuery object for the element being positioned.
-   * @param {jQuery} anchor - jQuery object for the element's anchor point.
-   * @param {String} position - a string relating to the desired position of the element, relative to it's anchor
-   * @param {Number} vOffset - integer pixel value of desired vertical separation between anchor and element.
-   * @param {Number} hOffset - integer pixel value of desired horizontal separation between anchor and element.
-   * @param {Boolean} isOverflow - if a collision event is detected, sets to true to default the element to full width - any desired offset.
-   * TODO alter/rewrite to work with `em` values as well/instead of pixels
-   */
-  function GetOffsets(element, anchor, position, vOffset, hOffset, isOverflow) {
-    var $eleDims = GetDimensions(element),
-        $anchorDims = anchor ? GetDimensions(anchor) : null;
-
-    switch (position) {
-      case 'top':
-        return {
-          left: Foundation.rtl() ? $anchorDims.offset.left - $eleDims.width + $anchorDims.width : $anchorDims.offset.left,
-          top: $anchorDims.offset.top - ($eleDims.height + vOffset)
-        };
-        break;
-      case 'left':
-        return {
-          left: $anchorDims.offset.left - ($eleDims.width + hOffset),
-          top: $anchorDims.offset.top
-        };
-        break;
-      case 'right':
-        return {
-          left: $anchorDims.offset.left + $anchorDims.width + hOffset,
-          top: $anchorDims.offset.top
-        };
-        break;
-      case 'center top':
-        return {
-          left: $anchorDims.offset.left + $anchorDims.width / 2 - $eleDims.width / 2,
-          top: $anchorDims.offset.top - ($eleDims.height + vOffset)
-        };
-        break;
-      case 'center bottom':
-        return {
-          left: isOverflow ? hOffset : $anchorDims.offset.left + $anchorDims.width / 2 - $eleDims.width / 2,
-          top: $anchorDims.offset.top + $anchorDims.height + vOffset
-        };
-        break;
-      case 'center left':
-        return {
-          left: $anchorDims.offset.left - ($eleDims.width + hOffset),
-          top: $anchorDims.offset.top + $anchorDims.height / 2 - $eleDims.height / 2
-        };
-        break;
-      case 'center right':
-        return {
-          left: $anchorDims.offset.left + $anchorDims.width + hOffset + 1,
-          top: $anchorDims.offset.top + $anchorDims.height / 2 - $eleDims.height / 2
-        };
-        break;
-      case 'center':
-        return {
-          left: $eleDims.windowDims.offset.left + $eleDims.windowDims.width / 2 - $eleDims.width / 2,
-          top: $eleDims.windowDims.offset.top + $eleDims.windowDims.height / 2 - $eleDims.height / 2
-        };
-        break;
-      case 'reveal':
-        return {
-          left: ($eleDims.windowDims.width - $eleDims.width) / 2,
-          top: $eleDims.windowDims.offset.top + vOffset
-        };
-      case 'reveal full':
-        return {
-          left: $eleDims.windowDims.offset.left,
-          top: $eleDims.windowDims.offset.top
-        };
-        break;
-      case 'left bottom':
-        return {
-          left: $anchorDims.offset.left,
-          top: $anchorDims.offset.top + $anchorDims.height + vOffset
-        };
-        break;
-      case 'right bottom':
-        return {
-          left: $anchorDims.offset.left + $anchorDims.width + hOffset - $eleDims.width,
-          top: $anchorDims.offset.top + $anchorDims.height + vOffset
-        };
-        break;
-      default:
-        return {
-          left: Foundation.rtl() ? $anchorDims.offset.left - $eleDims.width + $anchorDims.width : $anchorDims.offset.left + hOffset,
-          top: $anchorDims.offset.top + $anchorDims.height + vOffset
-        };
-    }
-  }
-}(jQuery);
-/*******************************************
- *                                         *
- * This util was created by Marius Olbertz *
- * Please thank Marius on GitHub /owlbertz *
- * or the web http://www.mariusolbertz.de/ *
- *                                         *
- ******************************************/
-
-'use strict';
-
-!function ($) {
-
-  var keyCodes = {
-    9: 'TAB',
-    13: 'ENTER',
-    27: 'ESCAPE',
-    32: 'SPACE',
-    37: 'ARROW_LEFT',
-    38: 'ARROW_UP',
-    39: 'ARROW_RIGHT',
-    40: 'ARROW_DOWN'
-  };
-
-  var commands = {};
-
-  var Keyboard = {
-    keys: getKeyCodes(keyCodes),
-
-    /**
-     * Parses the (keyboard) event and returns a String that represents its key
-     * Can be used like Foundation.parseKey(event) === Foundation.keys.SPACE
-     * @param {Event} event - the event generated by the event handler
-     * @return String key - String that represents the key pressed
-     */
-    parseKey: function (event) {
-      var key = keyCodes[event.which || event.keyCode] || String.fromCharCode(event.which).toUpperCase();
-
-      // Remove un-printable characters, e.g. for `fromCharCode` calls for CTRL only events
-      key = key.replace(/\W+/, '');
-
-      if (event.shiftKey) key = 'SHIFT_' + key;
-      if (event.ctrlKey) key = 'CTRL_' + key;
-      if (event.altKey) key = 'ALT_' + key;
-
-      // Remove trailing underscore, in case only modifiers were used (e.g. only `CTRL_ALT`)
-      key = key.replace(/_$/, '');
-
-      return key;
-    },
-
-
-    /**
-     * Handles the given (keyboard) event
-     * @param {Event} event - the event generated by the event handler
-     * @param {String} component - Foundation component's name, e.g. Slider or Reveal
-     * @param {Objects} functions - collection of functions that are to be executed
-     */
-    handleKey: function (event, component, functions) {
-      var commandList = commands[component],
-          keyCode = this.parseKey(event),
-          cmds,
-          command,
-          fn;
-
-      if (!commandList) return console.warn('Component not defined!');
-
-      if (typeof commandList.ltr === 'undefined') {
-        // this component does not differentiate between ltr and rtl
-        cmds = commandList; // use plain list
-      } else {
-        // merge ltr and rtl: if document is rtl, rtl overwrites ltr and vice versa
-        if (Foundation.rtl()) cmds = $.extend({}, commandList.ltr, commandList.rtl);else cmds = $.extend({}, commandList.rtl, commandList.ltr);
-      }
-      command = cmds[keyCode];
-
-      fn = functions[command];
-      if (fn && typeof fn === 'function') {
-        // execute function  if exists
-        var returnValue = fn.apply();
-        if (functions.handled || typeof functions.handled === 'function') {
-          // execute function when event was handled
-          functions.handled(returnValue);
-        }
-      } else {
-        if (functions.unhandled || typeof functions.unhandled === 'function') {
-          // execute function when event was not handled
-          functions.unhandled();
-        }
-      }
-    },
-
-
-    /**
-     * Finds all focusable elements within the given `$element`
-     * @param {jQuery} $element - jQuery object to search within
-     * @return {jQuery} $focusable - all focusable elements within `$element`
-     */
-    findFocusable: function ($element) {
-      if (!$element) {
-        return false;
-      }
-      return $element.find('a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), iframe, object, embed, *[tabindex], *[contenteditable]').filter(function () {
-        if (!$(this).is(':visible') || $(this).attr('tabindex') < 0) {
-          return false;
-        } //only have visible elements and those that have a tabindex greater or equal 0
-        return true;
-      });
-    },
-
-
-    /**
-     * Returns the component name name
-     * @param {Object} component - Foundation component, e.g. Slider or Reveal
-     * @return String componentName
-     */
-
-    register: function (componentName, cmds) {
-      commands[componentName] = cmds;
-    },
-
-
-    /**
-     * Traps the focus in the given element.
-     * @param  {jQuery} $element  jQuery object to trap the foucs into.
-     */
-    trapFocus: function ($element) {
-      var $focusable = Foundation.Keyboard.findFocusable($element),
-          $firstFocusable = $focusable.eq(0),
-          $lastFocusable = $focusable.eq(-1);
-
-      $element.on('keydown.zf.trapfocus', function (event) {
-        if (event.target === $lastFocusable[0] && Foundation.Keyboard.parseKey(event) === 'TAB') {
-          event.preventDefault();
-          $firstFocusable.focus();
-        } else if (event.target === $firstFocusable[0] && Foundation.Keyboard.parseKey(event) === 'SHIFT_TAB') {
-          event.preventDefault();
-          $lastFocusable.focus();
-        }
-      });
-    },
-
-    /**
-     * Releases the trapped focus from the given element.
-     * @param  {jQuery} $element  jQuery object to release the focus for.
-     */
-    releaseFocus: function ($element) {
-      $element.off('keydown.zf.trapfocus');
-    }
-  };
-
-  /*
-   * Constants for easier comparing.
-   * Can be used like Foundation.parseKey(event) === Foundation.keys.SPACE
-   */
-  function getKeyCodes(kcs) {
-    var k = {};
-    for (var kc in kcs) {
-      k[kcs[kc]] = kcs[kc];
-    }return k;
-  }
-
-  Foundation.Keyboard = Keyboard;
-}(jQuery);
-'use strict';
-
-!function ($) {
-
-  // Default set of media queries
-  var defaultQueries = {
-    'default': 'only screen',
-    landscape: 'only screen and (orientation: landscape)',
-    portrait: 'only screen and (orientation: portrait)',
-    retina: 'only screen and (-webkit-min-device-pixel-ratio: 2),' + 'only screen and (min--moz-device-pixel-ratio: 2),' + 'only screen and (-o-min-device-pixel-ratio: 2/1),' + 'only screen and (min-device-pixel-ratio: 2),' + 'only screen and (min-resolution: 192dpi),' + 'only screen and (min-resolution: 2dppx)'
-  };
-
-  var MediaQuery = {
-    queries: [],
-
-    current: '',
-
-    /**
-     * Initializes the media query helper, by extracting the breakpoint list from the CSS and activating the breakpoint watcher.
-     * @function
-     * @private
-     */
-    _init: function () {
-      var self = this;
-      var extractedStyles = $('.foundation-mq').css('font-family');
-      var namedQueries;
-
-      namedQueries = parseStyleToObject(extractedStyles);
-
-      for (var key in namedQueries) {
-        if (namedQueries.hasOwnProperty(key)) {
-          self.queries.push({
-            name: key,
-            value: 'only screen and (min-width: ' + namedQueries[key] + ')'
-          });
-        }
-      }
-
-      this.current = this._getCurrentSize();
-
-      this._watcher();
-    },
-
-
-    /**
-     * Checks if the screen is at least as wide as a breakpoint.
-     * @function
-     * @param {String} size - Name of the breakpoint to check.
-     * @returns {Boolean} `true` if the breakpoint matches, `false` if it's smaller.
-     */
-    atLeast: function (size) {
-      var query = this.get(size);
-
-      if (query) {
-        return window.matchMedia(query).matches;
-      }
-
-      return false;
-    },
-
-
-    /**
-     * Checks if the screen matches to a breakpoint.
-     * @function
-     * @param {String} size - Name of the breakpoint to check, either 'small only' or 'small'. Omitting 'only' falls back to using atLeast() method.
-     * @returns {Boolean} `true` if the breakpoint matches, `false` if it does not.
-     */
-    is: function (size) {
-      size = size.trim().split(' ');
-      if (size.length > 1 && size[1] === 'only') {
-        if (size[0] === this._getCurrentSize()) return true;
-      } else {
-        return this.atLeast(size[0]);
-      }
-      return false;
-    },
-
-
-    /**
-     * Gets the media query of a breakpoint.
-     * @function
-     * @param {String} size - Name of the breakpoint to get.
-     * @returns {String|null} - The media query of the breakpoint, or `null` if the breakpoint doesn't exist.
-     */
-    get: function (size) {
-      for (var i in this.queries) {
-        if (this.queries.hasOwnProperty(i)) {
-          var query = this.queries[i];
-          if (size === query.name) return query.value;
-        }
-      }
-
-      return null;
-    },
-
-
-    /**
-     * Gets the current breakpoint name by testing every breakpoint and returning the last one to match (the biggest one).
-     * @function
-     * @private
-     * @returns {String} Name of the current breakpoint.
-     */
-    _getCurrentSize: function () {
-      var matched;
-
-      for (var i = 0; i < this.queries.length; i++) {
-        var query = this.queries[i];
-
-        if (window.matchMedia(query.value).matches) {
-          matched = query;
-        }
-      }
-
-      if (typeof matched === 'object') {
-        return matched.name;
-      } else {
-        return matched;
-      }
-    },
-
-
-    /**
-     * Activates the breakpoint watcher, which fires an event on the window whenever the breakpoint changes.
-     * @function
-     * @private
-     */
-    _watcher: function () {
-      var _this = this;
-
-      $(window).on('resize.zf.mediaquery', function () {
-        var newSize = _this._getCurrentSize(),
-            currentSize = _this.current;
-
-        if (newSize !== currentSize) {
-          // Change the current media query
-          _this.current = newSize;
-
-          // Broadcast the media query change on the window
-          $(window).trigger('changed.zf.mediaquery', [newSize, currentSize]);
-        }
-      });
-    }
-  };
-
-  Foundation.MediaQuery = MediaQuery;
-
-  // matchMedia() polyfill - Test a CSS media type/query in JS.
-  // Authors & copyright (c) 2012: Scott Jehl, Paul Irish, Nicholas Zakas, David Knight. Dual MIT/BSD license
-  window.matchMedia || (window.matchMedia = function () {
-    'use strict';
-
-    // For browsers that support matchMedium api such as IE 9 and webkit
-
-    var styleMedia = window.styleMedia || window.media;
-
-    // For those that don't support matchMedium
-    if (!styleMedia) {
-      var style = document.createElement('style'),
-          script = document.getElementsByTagName('script')[0],
-          info = null;
-
-      style.type = 'text/css';
-      style.id = 'matchmediajs-test';
-
-      script && script.parentNode && script.parentNode.insertBefore(style, script);
-
-      // 'style.currentStyle' is used by IE <= 8 and 'window.getComputedStyle' for all other browsers
-      info = 'getComputedStyle' in window && window.getComputedStyle(style, null) || style.currentStyle;
-
-      styleMedia = {
-        matchMedium: function (media) {
-          var text = '@media ' + media + '{ #matchmediajs-test { width: 1px; } }';
-
-          // 'style.styleSheet' is used by IE <= 8 and 'style.textContent' for all other browsers
-          if (style.styleSheet) {
-            style.styleSheet.cssText = text;
-          } else {
-            style.textContent = text;
-          }
-
-          // Test if media query is true or false
-          return info.width === '1px';
-        }
-      };
-    }
-
-    return function (media) {
-      return {
-        matches: styleMedia.matchMedium(media || 'all'),
-        media: media || 'all'
-      };
-    };
-  }());
-
-  // Thank you: https://github.com/sindresorhus/query-string
-  function parseStyleToObject(str) {
-    var styleObject = {};
-
-    if (typeof str !== 'string') {
-      return styleObject;
-    }
-
-    str = str.trim().slice(1, -1); // browsers re-quote string style values
-
-    if (!str) {
-      return styleObject;
-    }
-
-    styleObject = str.split('&').reduce(function (ret, param) {
-      var parts = param.replace(/\+/g, ' ').split('=');
-      var key = parts[0];
-      var val = parts[1];
-      key = decodeURIComponent(key);
-
-      // missing `=` should be `null`:
-      // http://w3.org/TR/2012/WD-url-20120524/#collect-url-parameters
-      val = val === undefined ? null : decodeURIComponent(val);
-
-      if (!ret.hasOwnProperty(key)) {
-        ret[key] = val;
-      } else if (Array.isArray(ret[key])) {
-        ret[key].push(val);
-      } else {
-        ret[key] = [ret[key], val];
-      }
-      return ret;
-    }, {});
-
-    return styleObject;
-  }
-
-  Foundation.MediaQuery = MediaQuery;
-}(jQuery);
-'use strict';
-
-!function ($) {
-
-  /**
-   * Motion module.
-   * @module foundation.motion
-   */
-
-  var initClasses = ['mui-enter', 'mui-leave'];
-  var activeClasses = ['mui-enter-active', 'mui-leave-active'];
-
-  var Motion = {
-    animateIn: function (element, animation, cb) {
-      animate(true, element, animation, cb);
-    },
-
-    animateOut: function (element, animation, cb) {
-      animate(false, element, animation, cb);
-    }
-  };
-
-  function Move(duration, elem, fn) {
-    var anim,
-        prog,
-        start = null;
-    // console.log('called');
-
-    if (duration === 0) {
-      fn.apply(elem);
-      elem.trigger('finished.zf.animate', [elem]).triggerHandler('finished.zf.animate', [elem]);
-      return;
-    }
-
-    function move(ts) {
-      if (!start) start = ts;
-      // console.log(start, ts);
-      prog = ts - start;
-      fn.apply(elem);
-
-      if (prog < duration) {
-        anim = window.requestAnimationFrame(move, elem);
-      } else {
-        window.cancelAnimationFrame(anim);
-        elem.trigger('finished.zf.animate', [elem]).triggerHandler('finished.zf.animate', [elem]);
-      }
-    }
-    anim = window.requestAnimationFrame(move);
-  }
-
-  /**
-   * Animates an element in or out using a CSS transition class.
-   * @function
-   * @private
-   * @param {Boolean} isIn - Defines if the animation is in or out.
-   * @param {Object} element - jQuery or HTML object to animate.
-   * @param {String} animation - CSS class to use.
-   * @param {Function} cb - Callback to run when animation is finished.
-   */
-  function animate(isIn, element, animation, cb) {
-    element = $(element).eq(0);
-
-    if (!element.length) return;
-
-    var initClass = isIn ? initClasses[0] : initClasses[1];
-    var activeClass = isIn ? activeClasses[0] : activeClasses[1];
-
-    // Set up the animation
-    reset();
-
-    element.addClass(animation).css('transition', 'none');
-
-    requestAnimationFrame(function () {
-      element.addClass(initClass);
-      if (isIn) element.show();
-    });
-
-    // Start the animation
-    requestAnimationFrame(function () {
-      element[0].offsetWidth;
-      element.css('transition', '').addClass(activeClass);
-    });
-
-    // Clean up the animation when it finishes
-    element.one(Foundation.transitionend(element), finish);
-
-    // Hides the element (for out animations), resets the element, and runs a callback
-    function finish() {
-      if (!isIn) element.hide();
-      reset();
-      if (cb) cb.apply(element);
-    }
-
-    // Resets transitions and removes motion-specific classes
-    function reset() {
-      element[0].style.transitionDuration = 0;
-      element.removeClass(initClass + ' ' + activeClass + ' ' + animation);
-    }
-  }
-
-  Foundation.Move = Move;
-  Foundation.Motion = Motion;
-}(jQuery);
-'use strict';
-
-!function ($) {
-
-  var Nest = {
-    Feather: function (menu) {
-      var type = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 'zf';
-
-      menu.attr('role', 'menubar');
-
-      var items = menu.find('li').attr({ 'role': 'menuitem' }),
-          subMenuClass = 'is-' + type + '-submenu',
-          subItemClass = subMenuClass + '-item',
-          hasSubClass = 'is-' + type + '-submenu-parent';
-
-      items.each(function () {
-        var $item = $(this),
-            $sub = $item.children('ul');
-
-        if ($sub.length) {
-          $item.addClass(hasSubClass).attr({
-            'aria-haspopup': true,
-            'aria-label': $item.children('a:first').text()
-          });
-          // Note:  Drilldowns behave differently in how they hide, and so need
-          // additional attributes.  We should look if this possibly over-generalized
-          // utility (Nest) is appropriate when we rework menus in 6.4
-          if (type === 'drilldown') {
-            $item.attr({ 'aria-expanded': false });
-          }
-
-          $sub.addClass('submenu ' + subMenuClass).attr({
-            'data-submenu': '',
-            'role': 'menu'
-          });
-          if (type === 'drilldown') {
-            $sub.attr({ 'aria-hidden': true });
-          }
-        }
-
-        if ($item.parent('[data-submenu]').length) {
-          $item.addClass('is-submenu-item ' + subItemClass);
-        }
-      });
-
-      return;
-    },
-    Burn: function (menu, type) {
-      var //items = menu.find('li'),
-      subMenuClass = 'is-' + type + '-submenu',
-          subItemClass = subMenuClass + '-item',
-          hasSubClass = 'is-' + type + '-submenu-parent';
-
-      menu.find('>li, .menu, .menu > li').removeClass(subMenuClass + ' ' + subItemClass + ' ' + hasSubClass + ' is-submenu-item submenu is-active').removeAttr('data-submenu').css('display', '');
-
-      // console.log(      menu.find('.' + subMenuClass + ', .' + subItemClass + ', .has-submenu, .is-submenu-item, .submenu, [data-submenu]')
-      //           .removeClass(subMenuClass + ' ' + subItemClass + ' has-submenu is-submenu-item submenu')
-      //           .removeAttr('data-submenu'));
-      // items.each(function(){
-      //   var $item = $(this),
-      //       $sub = $item.children('ul');
-      //   if($item.parent('[data-submenu]').length){
-      //     $item.removeClass('is-submenu-item ' + subItemClass);
-      //   }
-      //   if($sub.length){
-      //     $item.removeClass('has-submenu');
-      //     $sub.removeClass('submenu ' + subMenuClass).removeAttr('data-submenu');
-      //   }
-      // });
-    }
-  };
-
-  Foundation.Nest = Nest;
-}(jQuery);
-'use strict';
-
-!function ($) {
-
-  function Timer(elem, options, cb) {
-    var _this = this,
-        duration = options.duration,
-        //options is an object for easily adding features later.
-    nameSpace = Object.keys(elem.data())[0] || 'timer',
-        remain = -1,
-        start,
-        timer;
-
-    this.isPaused = false;
-
-    this.restart = function () {
-      remain = -1;
-      clearTimeout(timer);
-      this.start();
-    };
-
-    this.start = function () {
-      this.isPaused = false;
-      // if(!elem.data('paused')){ return false; }//maybe implement this sanity check if used for other things.
-      clearTimeout(timer);
-      remain = remain <= 0 ? duration : remain;
-      elem.data('paused', false);
-      start = Date.now();
-      timer = setTimeout(function () {
-        if (options.infinite) {
-          _this.restart(); //rerun the timer.
-        }
-        if (cb && typeof cb === 'function') {
-          cb();
-        }
-      }, remain);
-      elem.trigger('timerstart.zf.' + nameSpace);
-    };
-
-    this.pause = function () {
-      this.isPaused = true;
-      //if(elem.data('paused')){ return false; }//maybe implement this sanity check if used for other things.
-      clearTimeout(timer);
-      elem.data('paused', true);
-      var end = Date.now();
-      remain = remain - (end - start);
-      elem.trigger('timerpaused.zf.' + nameSpace);
-    };
-  }
-
-  /**
-   * Runs a callback function when images are fully loaded.
-   * @param {Object} images - Image(s) to check if loaded.
-   * @param {Func} callback - Function to execute when image is fully loaded.
-   */
-  function onImagesLoaded(images, callback) {
-    var self = this,
-        unloaded = images.length;
-
-    if (unloaded === 0) {
-      callback();
-    }
-
-    images.each(function () {
-      // Check if image is loaded
-      if (this.complete || this.readyState === 4 || this.readyState === 'complete') {
-        singleImageLoaded();
-      }
-      // Force load the image
-      else {
-          // fix for IE. See https://css-tricks.com/snippets/jquery/fixing-load-in-ie-for-cached-images/
-          var src = $(this).attr('src');
-          $(this).attr('src', src + (src.indexOf('?') >= 0 ? '&' : '?') + new Date().getTime());
-          $(this).one('load', function () {
-            singleImageLoaded();
-          });
-        }
-    });
-
-    function singleImageLoaded() {
-      unloaded--;
-      if (unloaded === 0) {
-        callback();
-      }
-    }
-  }
-
-  Foundation.Timer = Timer;
-  Foundation.onImagesLoaded = onImagesLoaded;
-}(jQuery);
-//**************************************************
-//**Work inspired by multiple jquery swipe plugins**
-//**Done by Yohai Ararat ***************************
-//**************************************************
-(function ($) {
-
-	$.spotSwipe = {
-		version: '1.0.0',
-		enabled: 'ontouchstart' in document.documentElement,
-		preventDefault: false,
-		moveThreshold: 75,
-		timeThreshold: 200
-	};
-
-	var startPosX,
-	    startPosY,
-	    startTime,
-	    elapsedTime,
-	    isMoving = false;
-
-	function onTouchEnd() {
-		//  alert(this);
-		this.removeEventListener('touchmove', onTouchMove);
-		this.removeEventListener('touchend', onTouchEnd);
-		isMoving = false;
-	}
-
-	function onTouchMove(e) {
-		if ($.spotSwipe.preventDefault) {
-			e.preventDefault();
-		}
-		if (isMoving) {
-			var x = e.touches[0].pageX;
-			var y = e.touches[0].pageY;
-			var dx = startPosX - x;
-			var dy = startPosY - y;
-			var dir;
-			elapsedTime = new Date().getTime() - startTime;
-			if (Math.abs(dx) >= $.spotSwipe.moveThreshold && elapsedTime <= $.spotSwipe.timeThreshold) {
-				dir = dx > 0 ? 'left' : 'right';
-			}
-			// else if(Math.abs(dy) >= $.spotSwipe.moveThreshold && elapsedTime <= $.spotSwipe.timeThreshold) {
-			//   dir = dy > 0 ? 'down' : 'up';
-			// }
-			if (dir) {
-				e.preventDefault();
-				onTouchEnd.call(this);
-				$(this).trigger('swipe', dir).trigger('swipe' + dir);
-			}
-		}
-	}
-
-	function onTouchStart(e) {
-		if (e.touches.length == 1) {
-			startPosX = e.touches[0].pageX;
-			startPosY = e.touches[0].pageY;
-			isMoving = true;
-			startTime = new Date().getTime();
-			this.addEventListener('touchmove', onTouchMove, false);
-			this.addEventListener('touchend', onTouchEnd, false);
-		}
-	}
-
-	function init() {
-		this.addEventListener && this.addEventListener('touchstart', onTouchStart, false);
-	}
-
-	function teardown() {
-		this.removeEventListener('touchstart', onTouchStart);
-	}
-
-	$.event.special.swipe = { setup: init };
-
-	$.each(['left', 'up', 'down', 'right'], function () {
-		$.event.special['swipe' + this] = { setup: function () {
-				$(this).on('swipe', $.noop);
-			} };
-	});
-})(jQuery);
-/****************************************************
- * Method for adding psuedo drag events to elements *
- ***************************************************/
-!function ($) {
-	$.fn.addTouch = function () {
-		this.each(function (i, el) {
-			$(el).bind('touchstart touchmove touchend touchcancel', function () {
-				//we pass the original event object because the jQuery event
-				//object is normalized to w3c specs and does not provide the TouchList
-				handleTouch(event);
-			});
-		});
-
-		var handleTouch = function (event) {
-			var touches = event.changedTouches,
-			    first = touches[0],
-			    eventTypes = {
-				touchstart: 'mousedown',
-				touchmove: 'mousemove',
-				touchend: 'mouseup'
-			},
-			    type = eventTypes[event.type],
-			    simulatedEvent;
-
-			if ('MouseEvent' in window && typeof window.MouseEvent === 'function') {
-				simulatedEvent = new window.MouseEvent(type, {
-					'bubbles': true,
-					'cancelable': true,
-					'screenX': first.screenX,
-					'screenY': first.screenY,
-					'clientX': first.clientX,
-					'clientY': first.clientY
-				});
-			} else {
-				simulatedEvent = document.createEvent('MouseEvent');
-				simulatedEvent.initMouseEvent(type, true, true, window, 1, first.screenX, first.screenY, first.clientX, first.clientY, false, false, false, false, 0 /*left*/, null);
-			}
-			first.target.dispatchEvent(simulatedEvent);
-		};
-	};
-}(jQuery);
-
-//**********************************
-//**From the jQuery Mobile Library**
-//**need to recreate functionality**
-//**and try to improve if possible**
-//**********************************
-
-/* Removing the jQuery function ****
-************************************
-
-(function( $, window, undefined ) {
-
-	var $document = $( document ),
-		// supportTouch = $.mobile.support.touch,
-		touchStartEvent = 'touchstart'//supportTouch ? "touchstart" : "mousedown",
-		touchStopEvent = 'touchend'//supportTouch ? "touchend" : "mouseup",
-		touchMoveEvent = 'touchmove'//supportTouch ? "touchmove" : "mousemove";
-
-	// setup new event shortcuts
-	$.each( ( "touchstart touchmove touchend " +
-		"swipe swipeleft swiperight" ).split( " " ), function( i, name ) {
-
-		$.fn[ name ] = function( fn ) {
-			return fn ? this.bind( name, fn ) : this.trigger( name );
-		};
-
-		// jQuery < 1.8
-		if ( $.attrFn ) {
-			$.attrFn[ name ] = true;
-		}
-	});
-
-	function triggerCustomEvent( obj, eventType, event, bubble ) {
-		var originalType = event.type;
-		event.type = eventType;
-		if ( bubble ) {
-			$.event.trigger( event, undefined, obj );
-		} else {
-			$.event.dispatch.call( obj, event );
-		}
-		event.type = originalType;
-	}
-
-	// also handles taphold
-
-	// Also handles swipeleft, swiperight
-	$.event.special.swipe = {
-
-		// More than this horizontal displacement, and we will suppress scrolling.
-		scrollSupressionThreshold: 30,
-
-		// More time than this, and it isn't a swipe.
-		durationThreshold: 1000,
-
-		// Swipe horizontal displacement must be more than this.
-		horizontalDistanceThreshold: window.devicePixelRatio >= 2 ? 15 : 30,
-
-		// Swipe vertical displacement must be less than this.
-		verticalDistanceThreshold: window.devicePixelRatio >= 2 ? 15 : 30,
-
-		getLocation: function ( event ) {
-			var winPageX = window.pageXOffset,
-				winPageY = window.pageYOffset,
-				x = event.clientX,
-				y = event.clientY;
-
-			if ( event.pageY === 0 && Math.floor( y ) > Math.floor( event.pageY ) ||
-				event.pageX === 0 && Math.floor( x ) > Math.floor( event.pageX ) ) {
-
-				// iOS4 clientX/clientY have the value that should have been
-				// in pageX/pageY. While pageX/page/ have the value 0
-				x = x - winPageX;
-				y = y - winPageY;
-			} else if ( y < ( event.pageY - winPageY) || x < ( event.pageX - winPageX ) ) {
-
-				// Some Android browsers have totally bogus values for clientX/Y
-				// when scrolling/zooming a page. Detectable since clientX/clientY
-				// should never be smaller than pageX/pageY minus page scroll
-				x = event.pageX - winPageX;
-				y = event.pageY - winPageY;
-			}
-
-			return {
-				x: x,
-				y: y
-			};
-		},
-
-		start: function( event ) {
-			var data = event.originalEvent.touches ?
-					event.originalEvent.touches[ 0 ] : event,
-				location = $.event.special.swipe.getLocation( data );
-			return {
-						time: ( new Date() ).getTime(),
-						coords: [ location.x, location.y ],
-						origin: $( event.target )
-					};
-		},
-
-		stop: function( event ) {
-			var data = event.originalEvent.touches ?
-					event.originalEvent.touches[ 0 ] : event,
-				location = $.event.special.swipe.getLocation( data );
-			return {
-						time: ( new Date() ).getTime(),
-						coords: [ location.x, location.y ]
-					};
-		},
-
-		handleSwipe: function( start, stop, thisObject, origTarget ) {
-			if ( stop.time - start.time < $.event.special.swipe.durationThreshold &&
-				Math.abs( start.coords[ 0 ] - stop.coords[ 0 ] ) > $.event.special.swipe.horizontalDistanceThreshold &&
-				Math.abs( start.coords[ 1 ] - stop.coords[ 1 ] ) < $.event.special.swipe.verticalDistanceThreshold ) {
-				var direction = start.coords[0] > stop.coords[ 0 ] ? "swipeleft" : "swiperight";
-
-				triggerCustomEvent( thisObject, "swipe", $.Event( "swipe", { target: origTarget, swipestart: start, swipestop: stop }), true );
-				triggerCustomEvent( thisObject, direction,$.Event( direction, { target: origTarget, swipestart: start, swipestop: stop } ), true );
-				return true;
-			}
-			return false;
-
-		},
-
-		// This serves as a flag to ensure that at most one swipe event event is
-		// in work at any given time
-		eventInProgress: false,
-
-		setup: function() {
-			var events,
-				thisObject = this,
-				$this = $( thisObject ),
-				context = {};
-
-			// Retrieve the events data for this element and add the swipe context
-			events = $.data( this, "mobile-events" );
-			if ( !events ) {
-				events = { length: 0 };
-				$.data( this, "mobile-events", events );
-			}
-			events.length++;
-			events.swipe = context;
-
-			context.start = function( event ) {
-
-				// Bail if we're already working on a swipe event
-				if ( $.event.special.swipe.eventInProgress ) {
-					return;
-				}
-				$.event.special.swipe.eventInProgress = true;
-
-				var stop,
-					start = $.event.special.swipe.start( event ),
-					origTarget = event.target,
-					emitted = false;
-
-				context.move = function( event ) {
-					if ( !start || event.isDefaultPrevented() ) {
-						return;
-					}
-
-					stop = $.event.special.swipe.stop( event );
-					if ( !emitted ) {
-						emitted = $.event.special.swipe.handleSwipe( start, stop, thisObject, origTarget );
-						if ( emitted ) {
-
-							// Reset the context to make way for the next swipe event
-							$.event.special.swipe.eventInProgress = false;
-						}
-					}
-					// prevent scrolling
-					if ( Math.abs( start.coords[ 0 ] - stop.coords[ 0 ] ) > $.event.special.swipe.scrollSupressionThreshold ) {
-						event.preventDefault();
-					}
-				};
-
-				context.stop = function() {
-						emitted = true;
-
-						// Reset the context to make way for the next swipe event
-						$.event.special.swipe.eventInProgress = false;
-						$document.off( touchMoveEvent, context.move );
-						context.move = null;
-				};
-
-				$document.on( touchMoveEvent, context.move )
-					.one( touchStopEvent, context.stop );
-			};
-			$this.on( touchStartEvent, context.start );
-		},
-
-		teardown: function() {
-			var events, context;
-
-			events = $.data( this, "mobile-events" );
-			if ( events ) {
-				context = events.swipe;
-				delete events.swipe;
-				events.length--;
-				if ( events.length === 0 ) {
-					$.removeData( this, "mobile-events" );
-				}
-			}
-
-			if ( context ) {
-				if ( context.start ) {
-					$( this ).off( touchStartEvent, context.start );
-				}
-				if ( context.move ) {
-					$document.off( touchMoveEvent, context.move );
-				}
-				if ( context.stop ) {
-					$document.off( touchStopEvent, context.stop );
-				}
-			}
-		}
-	};
-	$.each({
-		swipeleft: "swipe.left",
-		swiperight: "swipe.right"
-	}, function( event, sourceEvent ) {
-
-		$.event.special[ event ] = {
-			setup: function() {
-				$( this ).bind( sourceEvent, $.noop );
-			},
-			teardown: function() {
-				$( this ).unbind( sourceEvent );
-			}
-		};
-	});
-})( jQuery, this );
-*/
-'use strict';
-
-!function ($) {
-
-  var MutationObserver = function () {
-    var prefixes = ['WebKit', 'Moz', 'O', 'Ms', ''];
-    for (var i = 0; i < prefixes.length; i++) {
-      if (prefixes[i] + 'MutationObserver' in window) {
-        return window[prefixes[i] + 'MutationObserver'];
-      }
-    }
-    return false;
-  }();
-
-  var triggers = function (el, type) {
-    el.data(type).split(' ').forEach(function (id) {
-      $('#' + id)[type === 'close' ? 'trigger' : 'triggerHandler'](type + '.zf.trigger', [el]);
-    });
-  };
-  // Elements with [data-open] will reveal a plugin that supports it when clicked.
-  $(document).on('click.zf.trigger', '[data-open]', function () {
-    triggers($(this), 'open');
-  });
-
-  // Elements with [data-close] will close a plugin that supports it when clicked.
-  // If used without a value on [data-close], the event will bubble, allowing it to close a parent component.
-  $(document).on('click.zf.trigger', '[data-close]', function () {
-    var id = $(this).data('close');
-    if (id) {
-      triggers($(this), 'close');
-    } else {
-      $(this).trigger('close.zf.trigger');
-    }
-  });
-
-  // Elements with [data-toggle] will toggle a plugin that supports it when clicked.
-  $(document).on('click.zf.trigger', '[data-toggle]', function () {
-    var id = $(this).data('toggle');
-    if (id) {
-      triggers($(this), 'toggle');
-    } else {
-      $(this).trigger('toggle.zf.trigger');
-    }
-  });
-
-  // Elements with [data-closable] will respond to close.zf.trigger events.
-  $(document).on('close.zf.trigger', '[data-closable]', function (e) {
-    e.stopPropagation();
-    var animation = $(this).data('closable');
-
-    if (animation !== '') {
-      Foundation.Motion.animateOut($(this), animation, function () {
-        $(this).trigger('closed.zf');
-      });
-    } else {
-      $(this).fadeOut().trigger('closed.zf');
-    }
-  });
-
-  $(document).on('focus.zf.trigger blur.zf.trigger', '[data-toggle-focus]', function () {
-    var id = $(this).data('toggle-focus');
-    $('#' + id).triggerHandler('toggle.zf.trigger', [$(this)]);
-  });
-
-  /**
-  * Fires once after all other scripts have loaded
-  * @function
-  * @private
-  */
-  $(window).on('load', function () {
-    checkListeners();
-  });
-
-  function checkListeners() {
-    eventsListener();
-    resizeListener();
-    scrollListener();
-    mutateListener();
-    closemeListener();
-  }
-
-  //******** only fires this function once on load, if there's something to watch ********
-  function closemeListener(pluginName) {
-    var yetiBoxes = $('[data-yeti-box]'),
-        plugNames = ['dropdown', 'tooltip', 'reveal'];
-
-    if (pluginName) {
-      if (typeof pluginName === 'string') {
-        plugNames.push(pluginName);
-      } else if (typeof pluginName === 'object' && typeof pluginName[0] === 'string') {
-        plugNames.concat(pluginName);
-      } else {
-        console.error('Plugin names must be strings');
-      }
-    }
-    if (yetiBoxes.length) {
-      var listeners = plugNames.map(function (name) {
-        return 'closeme.zf.' + name;
-      }).join(' ');
-
-      $(window).off(listeners).on(listeners, function (e, pluginId) {
-        var plugin = e.namespace.split('.')[0];
-        var plugins = $('[data-' + plugin + ']').not('[data-yeti-box="' + pluginId + '"]');
-
-        plugins.each(function () {
-          var _this = $(this);
-
-          _this.triggerHandler('close.zf.trigger', [_this]);
-        });
-      });
-    }
-  }
-
-  function resizeListener(debounce) {
-    var timer = void 0,
-        $nodes = $('[data-resize]');
-    if ($nodes.length) {
-      $(window).off('resize.zf.trigger').on('resize.zf.trigger', function (e) {
-        if (timer) {
-          clearTimeout(timer);
-        }
-
-        timer = setTimeout(function () {
-
-          if (!MutationObserver) {
-            //fallback for IE 9
-            $nodes.each(function () {
-              $(this).triggerHandler('resizeme.zf.trigger');
-            });
-          }
-          //trigger all listening elements and signal a resize event
-          $nodes.attr('data-events', "resize");
-        }, debounce || 10); //default time to emit resize event
-      });
-    }
-  }
-
-  function scrollListener(debounce) {
-    var timer = void 0,
-        $nodes = $('[data-scroll]');
-    if ($nodes.length) {
-      $(window).off('scroll.zf.trigger').on('scroll.zf.trigger', function (e) {
-        if (timer) {
-          clearTimeout(timer);
-        }
-
-        timer = setTimeout(function () {
-
-          if (!MutationObserver) {
-            //fallback for IE 9
-            $nodes.each(function () {
-              $(this).triggerHandler('scrollme.zf.trigger');
-            });
-          }
-          //trigger all listening elements and signal a scroll event
-          $nodes.attr('data-events', "scroll");
-        }, debounce || 10); //default time to emit scroll event
-      });
-    }
-  }
-
-  function mutateListener(debounce) {
-    var $nodes = $('[data-mutate]');
-    if ($nodes.length && MutationObserver) {
-      //trigger all listening elements and signal a mutate event
-      //no IE 9 or 10
-      $nodes.each(function () {
-        $(this).triggerHandler('mutateme.zf.trigger');
-      });
-    }
-  }
-
-  function eventsListener() {
-    if (!MutationObserver) {
-      return false;
-    }
-    var nodes = document.querySelectorAll('[data-resize], [data-scroll], [data-mutate]');
-
-    //element callback
-    var listeningElementsMutation = function (mutationRecordsList) {
-      var $target = $(mutationRecordsList[0].target);
-
-      //trigger the event handler for the element depending on type
-      switch (mutationRecordsList[0].type) {
-
-        case "attributes":
-          if ($target.attr("data-events") === "scroll" && mutationRecordsList[0].attributeName === "data-events") {
-            $target.triggerHandler('scrollme.zf.trigger', [$target, window.pageYOffset]);
-          }
-          if ($target.attr("data-events") === "resize" && mutationRecordsList[0].attributeName === "data-events") {
-            $target.triggerHandler('resizeme.zf.trigger', [$target]);
-          }
-          if (mutationRecordsList[0].attributeName === "style") {
-            $target.closest("[data-mutate]").attr("data-events", "mutate");
-            $target.closest("[data-mutate]").triggerHandler('mutateme.zf.trigger', [$target.closest("[data-mutate]")]);
-          }
-          break;
-
-        case "childList":
-          $target.closest("[data-mutate]").attr("data-events", "mutate");
-          $target.closest("[data-mutate]").triggerHandler('mutateme.zf.trigger', [$target.closest("[data-mutate]")]);
-          break;
-
-        default:
-          return false;
-        //nothing
-      }
-    };
-
-    if (nodes.length) {
-      //for each element that needs to listen for resizing, scrolling, or mutation add a single observer
-      for (var i = 0; i <= nodes.length - 1; i++) {
-        var elementObserver = new MutationObserver(listeningElementsMutation);
-        elementObserver.observe(nodes[i], { attributes: true, childList: true, characterData: false, subtree: true, attributeFilter: ["data-events", "style"] });
-      }
-    }
-  }
-
-  // ------------------------------------
-
-  // [PH]
-  // Foundation.CheckWatchers = checkWatchers;
-  Foundation.IHearYou = checkListeners;
-  // Foundation.ISeeYou = scrollListener;
-  // Foundation.IFeelYou = closemeListener;
-}(jQuery);
-
-// function domMutationObserver(debounce) {
-//   // !!! This is coming soon and needs more work; not active  !!! //
-//   var timer,
-//   nodes = document.querySelectorAll('[data-mutate]');
-//   //
-//   if (nodes.length) {
-//     // var MutationObserver = (function () {
-//     //   var prefixes = ['WebKit', 'Moz', 'O', 'Ms', ''];
-//     //   for (var i=0; i < prefixes.length; i++) {
-//     //     if (prefixes[i] + 'MutationObserver' in window) {
-//     //       return window[prefixes[i] + 'MutationObserver'];
-//     //     }
-//     //   }
-//     //   return false;
-//     // }());
-//
-//
-//     //for the body, we need to listen for all changes effecting the style and class attributes
-//     var bodyObserver = new MutationObserver(bodyMutation);
-//     bodyObserver.observe(document.body, { attributes: true, childList: true, characterData: false, subtree:true, attributeFilter:["style", "class"]});
-//
-//
-//     //body callback
-//     function bodyMutation(mutate) {
-//       //trigger all listening elements and signal a mutation event
-//       if (timer) { clearTimeout(timer); }
-//
-//       timer = setTimeout(function() {
-//         bodyObserver.disconnect();
-//         $('[data-mutate]').attr('data-events',"mutate");
-//       }, debounce || 150);
-//     }
-//   }
-// }
-'use strict';
-
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-!function ($) {
-
-  /**
-   * Tooltip module.
-   * @module foundation.tooltip
-   * @requires foundation.util.box
-   * @requires foundation.util.mediaQuery
-   * @requires foundation.util.triggers
-   */
-
-  var Tooltip = function () {
-    /**
-     * Creates a new instance of a Tooltip.
-     * @class
-     * @fires Tooltip#init
-     * @param {jQuery} element - jQuery object to attach a tooltip to.
-     * @param {Object} options - object to extend the default configuration.
-     */
-    function Tooltip(element, options) {
-      _classCallCheck(this, Tooltip);
-
-      this.$element = element;
-      this.options = $.extend({}, Tooltip.defaults, this.$element.data(), options);
-
-      this.isActive = false;
-      this.isClick = false;
-      this._init();
-
-      Foundation.registerPlugin(this, 'Tooltip');
-    }
-
-    /**
-     * Initializes the tooltip by setting the creating the tip element, adding it's text, setting private variables and setting attributes on the anchor.
-     * @private
-     */
-
-
-    _createClass(Tooltip, [{
-      key: '_init',
-      value: function _init() {
-        var elemId = this.$element.attr('aria-describedby') || Foundation.GetYoDigits(6, 'tooltip');
-
-        this.options.positionClass = this.options.positionClass || this._getPositionClass(this.$element);
-        this.options.tipText = this.options.tipText || this.$element.attr('title');
-        this.template = this.options.template ? $(this.options.template) : this._buildTemplate(elemId);
-
-        if (this.options.allowHtml) {
-          this.template.appendTo(document.body).html(this.options.tipText).hide();
-        } else {
-          this.template.appendTo(document.body).text(this.options.tipText).hide();
-        }
-
-        this.$element.attr({
-          'title': '',
-          'aria-describedby': elemId,
-          'data-yeti-box': elemId,
-          'data-toggle': elemId,
-          'data-resize': elemId
-        }).addClass(this.options.triggerClass);
-
-        //helper variables to track movement on collisions
-        this.usedPositions = [];
-        this.counter = 4;
-        this.classChanged = false;
-
-        this._events();
-      }
-
-      /**
-       * Grabs the current positioning class, if present, and returns the value or an empty string.
-       * @private
-       */
-
-    }, {
-      key: '_getPositionClass',
-      value: function _getPositionClass(element) {
-        if (!element) {
-          return '';
-        }
-        // var position = element.attr('class').match(/top|left|right/g);
-        var position = element[0].className.match(/\b(top|left|right)\b/g);
-        position = position ? position[0] : '';
-        return position;
-      }
-    }, {
-      key: '_buildTemplate',
-
-      /**
-       * builds the tooltip element, adds attributes, and returns the template.
-       * @private
-       */
-      value: function _buildTemplate(id) {
-        var templateClasses = (this.options.tooltipClass + ' ' + this.options.positionClass + ' ' + this.options.templateClasses).trim();
-        var $template = $('<div></div>').addClass(templateClasses).attr({
-          'role': 'tooltip',
-          'aria-hidden': true,
-          'data-is-active': false,
-          'data-is-focus': false,
-          'id': id
-        });
-        return $template;
-      }
-
-      /**
-       * Function that gets called if a collision event is detected.
-       * @param {String} position - positioning class to try
-       * @private
-       */
-
-    }, {
-      key: '_reposition',
-      value: function _reposition(position) {
-        this.usedPositions.push(position ? position : 'bottom');
-
-        //default, try switching to opposite side
-        if (!position && this.usedPositions.indexOf('top') < 0) {
-          this.template.addClass('top');
-        } else if (position === 'top' && this.usedPositions.indexOf('bottom') < 0) {
-          this.template.removeClass(position);
-        } else if (position === 'left' && this.usedPositions.indexOf('right') < 0) {
-          this.template.removeClass(position).addClass('right');
-        } else if (position === 'right' && this.usedPositions.indexOf('left') < 0) {
-          this.template.removeClass(position).addClass('left');
-        }
-
-        //if default change didn't work, try bottom or left first
-        else if (!position && this.usedPositions.indexOf('top') > -1 && this.usedPositions.indexOf('left') < 0) {
-            this.template.addClass('left');
-          } else if (position === 'top' && this.usedPositions.indexOf('bottom') > -1 && this.usedPositions.indexOf('left') < 0) {
-            this.template.removeClass(position).addClass('left');
-          } else if (position === 'left' && this.usedPositions.indexOf('right') > -1 && this.usedPositions.indexOf('bottom') < 0) {
-            this.template.removeClass(position);
-          } else if (position === 'right' && this.usedPositions.indexOf('left') > -1 && this.usedPositions.indexOf('bottom') < 0) {
-            this.template.removeClass(position);
-          }
-          //if nothing cleared, set to bottom
-          else {
-              this.template.removeClass(position);
-            }
-        this.classChanged = true;
-        this.counter--;
-      }
-
-      /**
-       * sets the position class of an element and recursively calls itself until there are no more possible positions to attempt, or the tooltip element is no longer colliding.
-       * if the tooltip is larger than the screen width, default to full width - any user selected margin
-       * @private
-       */
-
-    }, {
-      key: '_setPosition',
-      value: function _setPosition() {
-        var position = this._getPositionClass(this.template),
-            $tipDims = Foundation.Box.GetDimensions(this.template),
-            $anchorDims = Foundation.Box.GetDimensions(this.$element),
-            direction = position === 'left' ? 'left' : position === 'right' ? 'left' : 'top',
-            param = direction === 'top' ? 'height' : 'width',
-            offset = param === 'height' ? this.options.vOffset : this.options.hOffset,
-            _this = this;
-
-        if ($tipDims.width >= $tipDims.windowDims.width || !this.counter && !Foundation.Box.ImNotTouchingYou(this.template)) {
-          this.template.offset(Foundation.Box.GetOffsets(this.template, this.$element, 'center bottom', this.options.vOffset, this.options.hOffset, true)).css({
-            // this.$element.offset(Foundation.GetOffsets(this.template, this.$element, 'center bottom', this.options.vOffset, this.options.hOffset, true)).css({
-            'width': $anchorDims.windowDims.width - this.options.hOffset * 2,
-            'height': 'auto'
-          });
-          return false;
-        }
-
-        this.template.offset(Foundation.Box.GetOffsets(this.template, this.$element, 'center ' + (position || 'bottom'), this.options.vOffset, this.options.hOffset));
-
-        while (!Foundation.Box.ImNotTouchingYou(this.template) && this.counter) {
-          this._reposition(position);
-          this._setPosition();
-        }
-      }
-
-      /**
-       * reveals the tooltip, and fires an event to close any other open tooltips on the page
-       * @fires Tooltip#closeme
-       * @fires Tooltip#show
-       * @function
-       */
-
-    }, {
-      key: 'show',
-      value: function show() {
-        if (this.options.showOn !== 'all' && !Foundation.MediaQuery.is(this.options.showOn)) {
-          // console.error('The screen is too small to display this tooltip');
-          return false;
-        }
-
-        var _this = this;
-        this.template.css('visibility', 'hidden').show();
-        this._setPosition();
-
-        /**
-         * Fires to close all other open tooltips on the page
-         * @event Closeme#tooltip
-         */
-        this.$element.trigger('closeme.zf.tooltip', this.template.attr('id'));
-
-        this.template.attr({
-          'data-is-active': true,
-          'aria-hidden': false
-        });
-        _this.isActive = true;
-        // console.log(this.template);
-        this.template.stop().hide().css('visibility', '').fadeIn(this.options.fadeInDuration, function () {
-          //maybe do stuff?
-        });
-        /**
-         * Fires when the tooltip is shown
-         * @event Tooltip#show
-         */
-        this.$element.trigger('show.zf.tooltip');
-      }
-
-      /**
-       * Hides the current tooltip, and resets the positioning class if it was changed due to collision
-       * @fires Tooltip#hide
-       * @function
-       */
-
-    }, {
-      key: 'hide',
-      value: function hide() {
-        // console.log('hiding', this.$element.data('yeti-box'));
-        var _this = this;
-        this.template.stop().attr({
-          'aria-hidden': true,
-          'data-is-active': false
-        }).fadeOut(this.options.fadeOutDuration, function () {
-          _this.isActive = false;
-          _this.isClick = false;
-          if (_this.classChanged) {
-            _this.template.removeClass(_this._getPositionClass(_this.template)).addClass(_this.options.positionClass);
-
-            _this.usedPositions = [];
-            _this.counter = 4;
-            _this.classChanged = false;
-          }
-        });
-        /**
-         * fires when the tooltip is hidden
-         * @event Tooltip#hide
-         */
-        this.$element.trigger('hide.zf.tooltip');
-      }
-
-      /**
-       * adds event listeners for the tooltip and its anchor
-       * TODO combine some of the listeners like focus and mouseenter, etc.
-       * @private
-       */
-
-    }, {
-      key: '_events',
-      value: function _events() {
-        var _this = this;
-        var $template = this.template;
-        var isFocus = false;
-
-        if (!this.options.disableHover) {
-
-          this.$element.on('mouseenter.zf.tooltip', function (e) {
-            if (!_this.isActive) {
-              _this.timeout = setTimeout(function () {
-                _this.show();
-              }, _this.options.hoverDelay);
-            }
-          }).on('mouseleave.zf.tooltip', function (e) {
-            clearTimeout(_this.timeout);
-            if (!isFocus || _this.isClick && !_this.options.clickOpen) {
-              _this.hide();
-            }
-          });
-        }
-
-        if (this.options.clickOpen) {
-          this.$element.on('mousedown.zf.tooltip', function (e) {
-            e.stopImmediatePropagation();
-            if (_this.isClick) {
-              //_this.hide();
-              // _this.isClick = false;
-            } else {
-              _this.isClick = true;
-              if ((_this.options.disableHover || !_this.$element.attr('tabindex')) && !_this.isActive) {
-                _this.show();
-              }
-            }
-          });
-        } else {
-          this.$element.on('mousedown.zf.tooltip', function (e) {
-            e.stopImmediatePropagation();
-            _this.isClick = true;
-          });
-        }
-
-        if (!this.options.disableForTouch) {
-          this.$element.on('tap.zf.tooltip touchend.zf.tooltip', function (e) {
-            _this.isActive ? _this.hide() : _this.show();
-          });
-        }
-
-        this.$element.on({
-          // 'toggle.zf.trigger': this.toggle.bind(this),
-          // 'close.zf.trigger': this.hide.bind(this)
-          'close.zf.trigger': this.hide.bind(this)
-        });
-
-        this.$element.on('focus.zf.tooltip', function (e) {
-          isFocus = true;
-          if (_this.isClick) {
-            // If we're not showing open on clicks, we need to pretend a click-launched focus isn't
-            // a real focus, otherwise on hover and come back we get bad behavior
-            if (!_this.options.clickOpen) {
-              isFocus = false;
-            }
-            return false;
-          } else {
-            _this.show();
-          }
-        }).on('focusout.zf.tooltip', function (e) {
-          isFocus = false;
-          _this.isClick = false;
-          _this.hide();
-        }).on('resizeme.zf.trigger', function () {
-          if (_this.isActive) {
-            _this._setPosition();
-          }
-        });
-      }
-
-      /**
-       * adds a toggle method, in addition to the static show() & hide() functions
-       * @function
-       */
-
-    }, {
-      key: 'toggle',
-      value: function toggle() {
-        if (this.isActive) {
-          this.hide();
-        } else {
-          this.show();
-        }
-      }
-
-      /**
-       * Destroys an instance of tooltip, removes template element from the view.
-       * @function
-       */
-
-    }, {
-      key: 'destroy',
-      value: function destroy() {
-        this.$element.attr('title', this.template.text()).off('.zf.trigger .zf.tooltip').removeClass('has-tip top right left').removeAttr('aria-describedby aria-haspopup data-disable-hover data-resize data-toggle data-tooltip data-yeti-box');
-
-        this.template.remove();
-
-        Foundation.unregisterPlugin(this);
-      }
-    }]);
-
-    return Tooltip;
-  }();
-
-  Tooltip.defaults = {
-    disableForTouch: false,
-    /**
-     * Time, in ms, before a tooltip should open on hover.
-     * @option
-     * @type {number}
-     * @default 200
-     */
-    hoverDelay: 200,
-    /**
-     * Time, in ms, a tooltip should take to fade into view.
-     * @option
-     * @type {number}
-     * @default 150
-     */
-    fadeInDuration: 150,
-    /**
-     * Time, in ms, a tooltip should take to fade out of view.
-     * @option
-     * @type {number}
-     * @default 150
-     */
-    fadeOutDuration: 150,
-    /**
-     * Disables hover events from opening the tooltip if set to true
-     * @option
-     * @type {boolean}
-     * @default false
-     */
-    disableHover: false,
-    /**
-     * Optional addtional classes to apply to the tooltip template on init.
-     * @option
-     * @type {string}
-     * @default ''
-     */
-    templateClasses: '',
-    /**
-     * Non-optional class added to tooltip templates. Foundation default is 'tooltip'.
-     * @option
-     * @type {string}
-     * @default 'tooltip'
-     */
-    tooltipClass: 'tooltip',
-    /**
-     * Class applied to the tooltip anchor element.
-     * @option
-     * @type {string}
-     * @default 'has-tip'
-     */
-    triggerClass: 'has-tip',
-    /**
-     * Minimum breakpoint size at which to open the tooltip.
-     * @option
-     * @type {string}
-     * @default 'small'
-     */
-    showOn: 'small',
-    /**
-     * Custom template to be used to generate markup for tooltip.
-     * @option
-     * @type {string}
-     * @default ''
-     */
-    template: '',
-    /**
-     * Text displayed in the tooltip template on open.
-     * @option
-     * @type {string}
-     * @default ''
-     */
-    tipText: '',
-    touchCloseText: 'Tap to close.',
-    /**
-     * Allows the tooltip to remain open if triggered with a click or touch event.
-     * @option
-     * @type {boolean}
-     * @default true
-     */
-    clickOpen: true,
-    /**
-     * Additional positioning classes, set by the JS
-     * @option
-     * @type {string}
-     * @default ''
-     */
-    positionClass: '',
-    /**
-     * Distance, in pixels, the template should push away from the anchor on the Y axis.
-     * @option
-     * @type {number}
-     * @default 10
-     */
-    vOffset: 10,
-    /**
-     * Distance, in pixels, the template should push away from the anchor on the X axis, if aligned to a side.
-     * @option
-     * @type {number}
-     * @default 12
-     */
-    hOffset: 12,
-    /**
-    * Allow HTML in tooltip. Warning: If you are loading user-generated content into tooltips,
-    * allowing HTML may open yourself up to XSS attacks.
-    * @option
-    * @type {boolean}
-    * @default false
-    */
-    allowHtml: false
-  };
-
-  /**
-   * TODO utilize resize event trigger
-   */
-
-  // Window exports
-  Foundation.plugin(Tooltip, 'Tooltip');
-}(jQuery);
-'use strict';
-
-var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
-
-function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
-
-!function ($) {
-
-  /**
-   * Reveal module.
-   * @module foundation.reveal
-   * @requires foundation.util.keyboard
-   * @requires foundation.util.box
-   * @requires foundation.util.triggers
-   * @requires foundation.util.mediaQuery
-   * @requires foundation.util.motion if using animations
-   */
-
-  var Reveal = function () {
-    /**
-     * Creates a new instance of Reveal.
-     * @class
-     * @param {jQuery} element - jQuery object to use for the modal.
-     * @param {Object} options - optional parameters.
-     */
-    function Reveal(element, options) {
-      _classCallCheck(this, Reveal);
-
-      this.$element = element;
-      this.options = $.extend({}, Reveal.defaults, this.$element.data(), options);
-      this._init();
-
-      Foundation.registerPlugin(this, 'Reveal');
-      Foundation.Keyboard.register('Reveal', {
-        'ENTER': 'open',
-        'SPACE': 'open',
-        'ESCAPE': 'close'
-      });
-    }
-
-    /**
-     * Initializes the modal by adding the overlay and close buttons, (if selected).
-     * @private
-     */
-
-
-    _createClass(Reveal, [{
-      key: '_init',
-      value: function _init() {
-        this.id = this.$element.attr('id');
-        this.isActive = false;
-        this.cached = { mq: Foundation.MediaQuery.current };
-        this.isMobile = mobileSniff();
-
-        this.$anchor = $('[data-open="' + this.id + '"]').length ? $('[data-open="' + this.id + '"]') : $('[data-toggle="' + this.id + '"]');
-        this.$anchor.attr({
-          'aria-controls': this.id,
-          'aria-haspopup': true,
-          'tabindex': 0
-        });
-
-        if (this.options.fullScreen || this.$element.hasClass('full')) {
-          this.options.fullScreen = true;
-          this.options.overlay = false;
-        }
-        if (this.options.overlay && !this.$overlay) {
-          this.$overlay = this._makeOverlay(this.id);
-        }
-
-        this.$element.attr({
-          'role': 'dialog',
-          'aria-hidden': true,
-          'data-yeti-box': this.id,
-          'data-resize': this.id
-        });
-
-        if (this.$overlay) {
-          this.$element.detach().appendTo(this.$overlay);
-        } else {
-          this.$element.detach().appendTo($(this.options.appendTo));
-          this.$element.addClass('without-overlay');
-        }
-        this._events();
-        if (this.options.deepLink && window.location.hash === '#' + this.id) {
-          $(window).one('load.zf.reveal', this.open.bind(this));
-        }
-      }
-
-      /**
-       * Creates an overlay div to display behind the modal.
-       * @private
-       */
-
-    }, {
-      key: '_makeOverlay',
-      value: function _makeOverlay() {
-        return $('<div></div>').addClass('reveal-overlay').appendTo(this.options.appendTo);
-      }
-
-      /**
-       * Updates position of modal
-       * TODO:  Figure out if we actually need to cache these values or if it doesn't matter
-       * @private
-       */
-
-    }, {
-      key: '_updatePosition',
-      value: function _updatePosition() {
-        var width = this.$element.outerWidth();
-        var outerWidth = $(window).width();
-        var height = this.$element.outerHeight();
-        var outerHeight = $(window).height();
-        var left, top;
-        if (this.options.hOffset === 'auto') {
-          left = parseInt((outerWidth - width) / 2, 10);
-        } else {
-          left = parseInt(this.options.hOffset, 10);
-        }
-        if (this.options.vOffset === 'auto') {
-          if (height > outerHeight) {
-            top = parseInt(Math.min(100, outerHeight / 10), 10);
-          } else {
-            top = parseInt((outerHeight - height) / 4, 10);
-          }
-        } else {
-          top = parseInt(this.options.vOffset, 10);
-        }
-        this.$element.css({ top: top + 'px' });
-        // only worry about left if we don't have an overlay or we havea  horizontal offset,
-        // otherwise we're perfectly in the middle
-        if (!this.$overlay || this.options.hOffset !== 'auto') {
-          this.$element.css({ left: left + 'px' });
-          this.$element.css({ margin: '0px' });
-        }
-      }
-
-      /**
-       * Adds event handlers for the modal.
-       * @private
-       */
-
-    }, {
-      key: '_events',
-      value: function _events() {
-        var _this2 = this;
-
-        var _this = this;
-
-        this.$element.on({
-          'open.zf.trigger': this.open.bind(this),
-          'close.zf.trigger': function (event, $element) {
-            if (event.target === _this.$element[0] || $(event.target).parents('[data-closable]')[0] === $element) {
-              // only close reveal when it's explicitly called
-              return _this2.close.apply(_this2);
-            }
-          },
-          'toggle.zf.trigger': this.toggle.bind(this),
-          'resizeme.zf.trigger': function () {
-            _this._updatePosition();
-          }
-        });
-
-        if (this.$anchor.length) {
-          this.$anchor.on('keydown.zf.reveal', function (e) {
-            if (e.which === 13 || e.which === 32) {
-              e.stopPropagation();
-              e.preventDefault();
-              _this.open();
-            }
-          });
-        }
-
-        if (this.options.closeOnClick && this.options.overlay) {
-          this.$overlay.off('.zf.reveal').on('click.zf.reveal', function (e) {
-            if (e.target === _this.$element[0] || $.contains(_this.$element[0], e.target) || !$.contains(document, e.target)) {
-              return;
-            }
-            _this.close();
-          });
-        }
-        if (this.options.deepLink) {
-          $(window).on('popstate.zf.reveal:' + this.id, this._handleState.bind(this));
-        }
-      }
-
-      /**
-       * Handles modal methods on back/forward button clicks or any other event that triggers popstate.
-       * @private
-       */
-
-    }, {
-      key: '_handleState',
-      value: function _handleState(e) {
-        if (window.location.hash === '#' + this.id && !this.isActive) {
-          this.open();
-        } else {
-          this.close();
-        }
-      }
-
-      /**
-       * Opens the modal controlled by `this.$anchor`, and closes all others by default.
-       * @function
-       * @fires Reveal#closeme
-       * @fires Reveal#open
-       */
-
-    }, {
-      key: 'open',
-      value: function open() {
-        var _this3 = this;
-
-        if (this.options.deepLink) {
-          var hash = '#' + this.id;
-
-          if (window.history.pushState) {
-            window.history.pushState(null, null, hash);
-          } else {
-            window.location.hash = hash;
-          }
-        }
-
-        this.isActive = true;
-
-        // Make elements invisible, but remove display: none so we can get size and positioning
-        this.$element.css({ 'visibility': 'hidden' }).show().scrollTop(0);
-        if (this.options.overlay) {
-          this.$overlay.css({ 'visibility': 'hidden' }).show();
-        }
-
-        this._updatePosition();
-
-        this.$element.hide().css({ 'visibility': '' });
-
-        if (this.$overlay) {
-          this.$overlay.css({ 'visibility': '' }).hide();
-          if (this.$element.hasClass('fast')) {
-            this.$overlay.addClass('fast');
-          } else if (this.$element.hasClass('slow')) {
-            this.$overlay.addClass('slow');
-          }
-        }
-
-        if (!this.options.multipleOpened) {
-          /**
-           * Fires immediately before the modal opens.
-           * Closes any other modals that are currently open
-           * @event Reveal#closeme
-           */
-          this.$element.trigger('closeme.zf.reveal', this.id);
-        }
-
-        var _this = this;
-
-        function addRevealOpenClasses() {
-          if (_this.isMobile) {
-            if (!_this.originalScrollPos) {
-              _this.originalScrollPos = window.pageYOffset;
-            }
-            $('html, body').addClass('is-reveal-open');
-          } else {
-            $('body').addClass('is-reveal-open');
-          }
-        }
-        // Motion UI method of reveal
-        if (this.options.animationIn) {
-          var afterAnimation = function () {
-            _this.$element.attr({
-              'aria-hidden': false,
-              'tabindex': -1
-            }).focus();
-            addRevealOpenClasses();
-            Foundation.Keyboard.trapFocus(_this.$element);
-          };
-
-          if (this.options.overlay) {
-            Foundation.Motion.animateIn(this.$overlay, 'fade-in');
-          }
-          Foundation.Motion.animateIn(this.$element, this.options.animationIn, function () {
-            if (_this3.$element) {
-              // protect against object having been removed
-              _this3.focusableElements = Foundation.Keyboard.findFocusable(_this3.$element);
-              afterAnimation();
-            }
-          });
-        }
-        // jQuery method of reveal
-        else {
-            if (this.options.overlay) {
-              this.$overlay.show(0);
-            }
-            this.$element.show(this.options.showDelay);
-          }
-
-        // handle accessibility
-        this.$element.attr({
-          'aria-hidden': false,
-          'tabindex': -1
-        }).focus();
-        Foundation.Keyboard.trapFocus(this.$element);
-
-        /**
-         * Fires when the modal has successfully opened.
-         * @event Reveal#open
-         */
-        this.$element.trigger('open.zf.reveal');
-
-        addRevealOpenClasses();
-
-        setTimeout(function () {
-          _this3._extraHandlers();
-        }, 0);
-      }
-
-      /**
-       * Adds extra event handlers for the body and window if necessary.
-       * @private
-       */
-
-    }, {
-      key: '_extraHandlers',
-      value: function _extraHandlers() {
-        var _this = this;
-        if (!this.$element) {
-          return;
-        } // If we're in the middle of cleanup, don't freak out
-        this.focusableElements = Foundation.Keyboard.findFocusable(this.$element);
-
-        if (!this.options.overlay && this.options.closeOnClick && !this.options.fullScreen) {
-          $('body').on('click.zf.reveal', function (e) {
-            if (e.target === _this.$element[0] || $.contains(_this.$element[0], e.target) || !$.contains(document, e.target)) {
-              return;
-            }
-            _this.close();
-          });
-        }
-
-        if (this.options.closeOnEsc) {
-          $(window).on('keydown.zf.reveal', function (e) {
-            Foundation.Keyboard.handleKey(e, 'Reveal', {
-              close: function () {
-                if (_this.options.closeOnEsc) {
-                  _this.close();
-                  _this.$anchor.focus();
-                }
-              }
-            });
-          });
-        }
-
-        // lock focus within modal while tabbing
-        this.$element.on('keydown.zf.reveal', function (e) {
-          var $target = $(this);
-          // handle keyboard event with keyboard util
-          Foundation.Keyboard.handleKey(e, 'Reveal', {
-            open: function () {
-              if (_this.$element.find(':focus').is(_this.$element.find('[data-close]'))) {
-                setTimeout(function () {
-                  // set focus back to anchor if close button has been activated
-                  _this.$anchor.focus();
-                }, 1);
-              } else if ($target.is(_this.focusableElements)) {
-                // dont't trigger if acual element has focus (i.e. inputs, links, ...)
-                _this.open();
-              }
-            },
-            close: function () {
-              if (_this.options.closeOnEsc) {
-                _this.close();
-                _this.$anchor.focus();
-              }
-            },
-            handled: function (preventDefault) {
-              if (preventDefault) {
-                e.preventDefault();
-              }
-            }
-          });
-        });
-      }
-
-      /**
-       * Closes the modal.
-       * @function
-       * @fires Reveal#closed
-       */
-
-    }, {
-      key: 'close',
-      value: function close() {
-        if (!this.isActive || !this.$element.is(':visible')) {
-          return false;
-        }
-        var _this = this;
-
-        // Motion UI method of hiding
-        if (this.options.animationOut) {
-          if (this.options.overlay) {
-            Foundation.Motion.animateOut(this.$overlay, 'fade-out', finishUp);
-          } else {
-            finishUp();
-          }
-
-          Foundation.Motion.animateOut(this.$element, this.options.animationOut);
-        }
-        // jQuery method of hiding
-        else {
-            if (this.options.overlay) {
-              this.$overlay.hide(0, finishUp);
-            } else {
-              finishUp();
-            }
-
-            this.$element.hide(this.options.hideDelay);
-          }
-
-        // Conditionals to remove extra event listeners added on open
-        if (this.options.closeOnEsc) {
-          $(window).off('keydown.zf.reveal');
-        }
-
-        if (!this.options.overlay && this.options.closeOnClick) {
-          $('body').off('click.zf.reveal');
-        }
-
-        this.$element.off('keydown.zf.reveal');
-
-        function finishUp() {
-          if (_this.isMobile) {
-            $('html, body').removeClass('is-reveal-open');
-            if (_this.originalScrollPos) {
-              $('body').scrollTop(_this.originalScrollPos);
-              _this.originalScrollPos = null;
-            }
-          } else {
-            $('body').removeClass('is-reveal-open');
-          }
-
-          Foundation.Keyboard.releaseFocus(_this.$element);
-
-          _this.$element.attr('aria-hidden', true);
-
-          /**
-          * Fires when the modal is done closing.
-          * @event Reveal#closed
-          */
-          _this.$element.trigger('closed.zf.reveal');
-        }
-
-        /**
-        * Resets the modal content
-        * This prevents a running video to keep going in the background
-        */
-        if (this.options.resetOnClose) {
-          this.$element.html(this.$element.html());
-        }
-
-        this.isActive = false;
-        if (_this.options.deepLink) {
-          if (window.history.replaceState) {
-            window.history.replaceState('', document.title, window.location.href.replace('#' + this.id, ''));
-          } else {
-            window.location.hash = '';
-          }
-        }
-      }
-
-      /**
-       * Toggles the open/closed state of a modal.
-       * @function
-       */
-
-    }, {
-      key: 'toggle',
-      value: function toggle() {
-        if (this.isActive) {
-          this.close();
-        } else {
-          this.open();
-        }
-      }
-    }, {
-      key: 'destroy',
-
-
-      /**
-       * Destroys an instance of a modal.
-       * @function
-       */
-      value: function destroy() {
-        if (this.options.overlay) {
-          this.$element.appendTo($(this.options.appendTo)); // move $element outside of $overlay to prevent error unregisterPlugin()
-          this.$overlay.hide().off().remove();
-        }
-        this.$element.hide().off();
-        this.$anchor.off('.zf');
-        $(window).off('.zf.reveal:' + this.id);
-
-        Foundation.unregisterPlugin(this);
-      }
-    }]);
-
-    return Reveal;
-  }();
-
-  Reveal.defaults = {
-    /**
-     * Motion-UI class to use for animated elements. If none used, defaults to simple show/hide.
-     * @option
-     * @type {string}
-     * @default ''
-     */
-    animationIn: '',
-    /**
-     * Motion-UI class to use for animated elements. If none used, defaults to simple show/hide.
-     * @option
-     * @type {string}
-     * @default ''
-     */
-    animationOut: '',
-    /**
-     * Time, in ms, to delay the opening of a modal after a click if no animation used.
-     * @option
-     * @type {number}
-     * @default 0
-     */
-    showDelay: 0,
-    /**
-     * Time, in ms, to delay the closing of a modal after a click if no animation used.
-     * @option
-     * @type {number}
-     * @default 0
-     */
-    hideDelay: 0,
-    /**
-     * Allows a click on the body/overlay to close the modal.
-     * @option
-     * @type {boolean}
-     * @default true
-     */
-    closeOnClick: true,
-    /**
-     * Allows the modal to close if the user presses the `ESCAPE` key.
-     * @option
-     * @type {boolean}
-     * @default true
-     */
-    closeOnEsc: true,
-    /**
-     * If true, allows multiple modals to be displayed at once.
-     * @option
-     * @type {boolean}
-     * @default false
-     */
-    multipleOpened: false,
-    /**
-     * Distance, in pixels, the modal should push down from the top of the screen.
-     * @option
-     * @type {number|string}
-     * @default auto
-     */
-    vOffset: 'auto',
-    /**
-     * Distance, in pixels, the modal should push in from the side of the screen.
-     * @option
-     * @type {number|string}
-     * @default auto
-     */
-    hOffset: 'auto',
-    /**
-     * Allows the modal to be fullscreen, completely blocking out the rest of the view. JS checks for this as well.
-     * @option
-     * @type {boolean}
-     * @default false
-     */
-    fullScreen: false,
-    /**
-     * Percentage of screen height the modal should push up from the bottom of the view.
-     * @option
-     * @type {number}
-     * @default 10
-     */
-    btmOffsetPct: 10,
-    /**
-     * Allows the modal to generate an overlay div, which will cover the view when modal opens.
-     * @option
-     * @type {boolean}
-     * @default true
-     */
-    overlay: true,
-    /**
-     * Allows the modal to remove and reinject markup on close. Should be true if using video elements w/o using provider's api, otherwise, videos will continue to play in the background.
-     * @option
-     * @type {boolean}
-     * @default false
-     */
-    resetOnClose: false,
-    /**
-     * Allows the modal to alter the url on open/close, and allows the use of the `back` button to close modals. ALSO, allows a modal to auto-maniacally open on page load IF the hash === the modal's user-set id.
-     * @option
-     * @type {boolean}
-     * @default false
-     */
-    deepLink: false,
-    /**
-    * Allows the modal to append to custom div.
-    * @option
-    * @type {string}
-    * @default "body"
-    */
-    appendTo: "body"
-
-  };
-
-  // Window exports
-  Foundation.plugin(Reveal, 'Reveal');
-
-  function iPhoneSniff() {
-    return (/iP(ad|hone|od).*OS/.test(window.navigator.userAgent)
-    );
-  }
-
-  function androidSniff() {
-    return (/Android/.test(window.navigator.userAgent)
-    );
-  }
-
-  function mobileSniff() {
-    return iPhoneSniff() || androidSniff();
-  }
-}(jQuery);
-'use strict';
-/*******************************************
- *                                         *
- * This Ride was created by Marius Olbertz *
- * Please thank Marius on GitHub /owlbertz *
- * or the web http://www.mariusolbertz.de/ *
- *                                         *
- ******************************************/
-
-var _createClass = function () {
-  function defineProperties(target, props) {
-    for (var i = 0; i < props.length; i++) {
-      var descriptor = props[i];descriptor.enumerable = descriptor.enumerable || false;descriptor.configurable = true;if ("value" in descriptor) descriptor.writable = true;Object.defineProperty(target, descriptor.key, descriptor);
-    }
-  }return function (Constructor, protoProps, staticProps) {
-    if (protoProps) defineProperties(Constructor.prototype, protoProps);if (staticProps) defineProperties(Constructor, staticProps);return Constructor;
-  };
-}();
-
-function _classCallCheck(instance, Constructor) {
-  if (!(instance instanceof Constructor)) {
-    throw new TypeError("Cannot call a class as a function");
-  }
-}
-
-!function (root, factory) {
-  if (typeof define === 'function' && define.amd) {
-    // AMD
-    define(['jquery', 'foundation'], factory);
-  } else if (typeof exports === 'object') {
-    // Node, CommonJS-like
-    module.exports = factory(require('jquery'));
-  } else {
-    // Browser globals (root is window)
-    root.returnExports = factory(root.jQuery);
-  }
-}(this, function ($) {
-  /**
-   * Joyride module.
-   * @module foundation.joyride
-   * @requires foundation.util.keyboard
-   * @requires foundation.Tooltip
-   * @requires foundation.Reveal
-   */
-  var Joyride = function () {
-    /**
-     * Creates a new instance of Joyride.
-     * @class
-     * @param {jQuery} element - jQuery object (list) to be used as the structure.
-     * @param {Object} options - object to extend the default configuration.
-     */
-
-    function Joyride(element, options) {
-      _classCallCheck(this, Joyride);
-
-      this.$element = element;
-      this.options = $.extend({}, Joyride.defaults, this.$element.data(), options);
-      this._init();
-
-      Foundation.registerPlugin(this, 'Joyride');
-      Foundation.Keyboard.register('Joyride', {
-        'ltr': {
-          'ARROW_RIGHT': 'next',
-          'ARROW_LEFT': 'previous',
-          'ESCAPE': 'close'
-        },
-        'rtl': {
-          'ARROW_RIGHT': 'previous',
-          'ARROW_LEFT': 'next'
-        }
-      });
-    }
-
-    /**
-     * Initializes the joyride by rendering required markup
-     * @private
-     */
-
-    _createClass(Joyride, [{
-      key: '_init',
-      value: function _init() {
-        this.id = this.$element.attr('id') || Foundation.GetYoDigits(6, 'joyride');
-        this.current = 0;
-        this.$items = $([]); // initialize empty collection
-        this.structure = this._parseList();
-        this._render(this.structure);
-        this._events();
-
-        if (this.options.autostart) {
-          // var _this = this;
-          // console.log(_this);
-          // $(window).on('load',function() {
-          //     _this.start();
-          // });
-          this.start();
-        }
-      }
-
-      /**
-       * Parses the list of the instance, stored in $element.
-       * @private
-       * @return {Array} structure
-       */
-
-    }, {
-      key: '_parseList',
-      value: function _parseList() {
-        var structure = [];
-        this.$element.find('li').each(function (i) {
-          var item = $.extend({}, {
-            text: $(this).html(),
-            $target: $($(this).data('target')),
-            isModal: !!!$($(this).data('target')).length,
-            closable: Joyride.defaults.closable
-          }, $(this).data());
-          structure.push(item);
-        });
-        return structure;
-      }
-
-      /**
-       * Creates the markup for the items
-       * @private
-       * @param {Array} structure the joyride's structure from _parseList
-       * @return {Object} markup jQuery representation of the generated markup
-       */
-
-    }, {
-      key: '_render',
-      value: function _render(structure) {
-        for (var s in structure) {
-          var options = $.extend({}, this.options, structure[s]),
-
-          // if specifc item has config, this should overwrite global settings
-          $item;
-
-          if (options.$target.length) {
-            // target element exists, create tooltip
-            var tooltip = new Foundation.Tooltip(structure[s].$target, {
-              positionClass: options.position,
-              disableHover: true,
-              clickOpen: false,
-              tooltipClass: 'tooltip joyride',
-              triggerClass: '',
-              hOffset: this.options.hOffset,
-              vOffset: this.options.vOffset,
-              showOn: 'all' // to avoid conflicts with standalone version when MQs are not set up properly. Should be somewhat rewritten
-            });
-            this.structure[s].item = tooltip;
-            $item = tooltip.template;
-          } else {
-            // not target, create modal with Reveal
-            var modal = new Foundation.Reveal($('<div class="reveal joyride"/>').appendTo($('body')));
-            this.structure[s].item = modal;
-            $item = modal.$element;
-          }
-          $item.attr({
-            'data-index': s,
-            'data-joyride-for': structure[s].target
-          }).html(structure[s].text);
-          if (options.keyboardAccess) {
-            $item.attr('tabindex', '-1');
-          }
-
-          this.$items = this.$items.add($item);
-
-          // add buttons
-          if (structure[s].nextText || options.showNext && s < structure.length - 1 || structure[s].prevText || options.showPrev && s > 0) {
-            var $buttons = $('<div class="joyride-buttons"/>');
-            if (structure[s].prevText || options.showPrev && s > 0) {
-              $buttons.append($(this.options.prevButton).text(options.prevText));
-            }
-            if (structure[s].nextText || options.showNext && s < structure.length - 1) {
-              $buttons.append($(this.options.nextButton).text(options.nextText));
-            }
-            $item.append($buttons);
-          }
-
-          // add close button
-          if (options.closable) {
-            var $close = $(this.options.closeButton);
-            $close.find('.show-for-sr').text(this.options.closeText);
-            $item.prepend($close);
-          }
-        }
-      }
-
-      /**
-       * Shows the item with the given index
-       * @private
-       * @param {Number} index of the item to be displayed
-       */
-
-    }, {
-      key: '_showItem',
-      value: function _showItem(index) {
-        if (this.structure[index].isModal) {
-          this.structure[index].item.open();
-        } else {
-          this.structure[index].item.show();
-        }
-        // scroll target into view if target exists
-        if (this.structure[index].$target.length) {
-          $('html, body').stop().animate({
-            'scrollTop': Math.max(0, this.$items.eq(index).offset().top - this.options.scrollOffset)
-          }, this.options.scrollSpeed);
-        }
-        if (this.options.keyboardAccess) {
-          this.$items.eq(index).focus();
-        }
-        this.current = index;
-      }
-
-      /**
-       * Hides the item with the given index
-       * @private
-       * @param {Number} index of the item to be hidden
-       */
-
-    }, {
-      key: '_hideItem',
-      value: function _hideItem(index) {
-        if (this.structure[index].isModal) {
-          this.structure[index].item.close();
-        } else {
-          this.structure[index].item.hide();
-        }
-      }
-
-      /**
-       * Hides all items
-       * @private
-       */
-
-    }, {
-      key: '_hideAll',
-      value: function _hideAll() {
-        for (var s in this.structure) {
-          this._hideItem(s);
-        }
-      }
-
-      /**
-       * Shows the next item in the ride
-       * @private
-       */
-
-    }, {
-      key: 'showNext',
-      value: function showNext() {
-        this._hideItem(this.current);
-        this._showItem(this.current + 1);
-      }
-
-      /**
-       * Shows the previous item in the ride
-       * @private
-       */
-
-    }, {
-      key: 'showPrev',
-      value: function showPrev() {
-        this._hideItem(this.current);
-        this._showItem(this.current - 1);
-      }
-
-      /**
-       * Starts the ride
-       * @private
-       * @return {Number} index - the index where to start, 0 by default
-       */
-
-    }, {
-      key: 'start',
-      value: function start(index) {
-        var index = index || 0;
-        this._hideAll();
-        this._showItem(index);
-      }
-
-      /**
-       * Adds event handlers for the modal.
-       * @private
-       */
-
-    }, {
-      key: '_events',
-      value: function _events() {
-        var _this = this;
-        $('[data-joyride-start="#' + _this.id + '"]').click(function () {
-          _this.start();
-        });
-
-        this.$items.on('click.zf.joyride', '[data-joyride-next]', function (e) {
-          _this.showNext();
-        }).on('click.zf.joyride', '[data-joyride-prev]', function (e) {
-          _this.showPrev();
-        }).on('click.zf.joyride', '[data-joyride-close]', function (e) {
-          e.preventDefault();
-          if (_this.structure[_this.current].closable) {
-            _this._hideItem(_this.current);
-          }
-        }).on('keydown.zf.joyride', function (e) {
-          var $element = $(this);
-          Foundation.Keyboard.handleKey(e, 'Joyride', {
-            next: function () {
-              if ($element.data('index') < _this.structure.length - 1) {
-                _this.showNext();
-              }
-            },
-            previous: function () {
-              if ($element.data('index') > 0) {
-                _this.showPrev();
-              }
-            },
-            close: function () {
-              if (_this.structure[_this.current].closable) {
-                _this._hideItem(_this.current);
-              }
-            },
-            handled: function () {
-              e.preventDefault();
-            }
-          });
-        });
-      }
-
-      /**
-       * Destroys an instance of a Joyride.
-       * @fires Joyride#destroyed
-       */
-
-    }, {
-      key: 'destroy',
-      value: function destroy() {
-        this.$element.hide();
-
-        for (var s in this.structure) {
-          this.structure[s].item.destroy();
-        }
-
-        /**
-         * Fires when the plugin has been destroyed.
-         * @event Joyride#destroyed
-         */
-        this.$element.trigger('destroyed.zf.joyride');
-
-        Foundation.unregisterPlugin(this);
-      }
-    }]);
-
-    return Joyride;
-  }();
-
-  Joyride.defaults = {
-    /**
-     * Autostart the joyride on page load.
-     * @option
-     * @example false
-     */
-    autostart: false,
-    /**
-     * Speed with which the page scrolls to the next stop.
-     * @option
-     * @example 1000
-     */
-    scrollSpeed: 1000,
-    /**
-     * Enable navigation via keyboard.
-     * @option
-     * @example true
-     */
-    keyboardAccess: true,
-    /**
-     * If the joyride elements should be closable.
-     * @option
-     * @example true
-     */
-    closable: true,
-    /**
-     * Text for the next button.
-     * @option
-     * @example 'Next'
-     */
-    nextText: 'Next',
-    /**
-     * Text for the previous button.
-     * @option
-     * @example 'Previous'
-     */
-    prevText: 'Previous',
-    /**
-     * Text for the close button (for accessibility purposes).
-     * @option
-     * @example 'Close'
-     */
-    closeText: 'Close',
-    /**
-     * Whether to show next button.
-     * @option
-     * @example true
-     */
-    showNext: true,
-    /**
-     * Whether to show previous button.
-     * @option
-     * @example true
-     */
-    showPrev: false,
-    /**
-     * Vertical offset for tooltips (see tooltip plugin configuration).
-     * @option
-     * @example true
-     */
-    vOffset: 10,
-    /**
-     * Horizontal offset for tooltips (see tooltip plugin configuration).
-     * @option
-     * @example true
-     */
-    hOffset: 12,
-    /**
-     * Offset while scrolling the window.
-     * @option
-     * @example 50
-     */
-    scrollOffset: 50,
-    /**
-     * Position of the tooltips (see tooltip plugin configuration).
-     * @option
-     * @example true
-     */
-    position: 'top center',
-    /**
-     * Templates for the close button.
-     * @option
-     * @example '<a href="#close" class="close" data-joyride-close><span aria-hidden="true">&times</span><span class="show-for-sr"></span></a>'
-     */
-    closeButton: '<a href="#close" class="close" data-joyride-close><span aria-hidden="true">&times</span><span class="show-for-sr"></span></a>',
-    /**
-     * Templates for the next button.
-     * @option
-     * @example '<button class="button" data-joyride-next></button>'
-     */
-    nextButton: '<button class="button" data-joyride-next></button>',
-    /**
-     * Templates for the pevious button.
-     * @option
-     * @example '<button class="button" data-joyride-prev></button>'
-     */
-    prevButton: '<button class="button" data-joyride-prev></button>'
-  };
-
-  // Window exports
-  Foundation.plugin(Joyride, 'Joyride');
-
-  return Joyride;
-});
